Example data with binary outcomes

A simulated data set containing 7994 observations, 5 continuous covariates and 100 providers.

Usage

data(ExampleDataBinary)

Format

A list containing the following elements:

Y

a vector representing binary outcomes with 0 or 1. Generated from a Bernoulli distribution with the probability of success (\(\boldsymbol{\mu}\)), which is determined by applying the logistic function to the linear combination of provider effects and covariates.

ID

a vector representing identifier for each provider. The number of individuals per provider is generated from a Poisson distribution with mean 80, with a minimum value of 11.

Z

a data frame containing 5 continuous variables. Generated from a multivariate normal distribution, where the mean is calculated as \((\gamma_i - \mu_{\gamma}) \cdot \rho / \sigma_{\gamma}\) with \(\mu_{\gamma} = \log(4/11)\) and \(\sigma_{\gamma} = 0.4\), and the correlation between covariates and provider effects is 0.1.

Examples

data(ExampleDataBinary)
head(ExampleDataBinary$Y)
#> [1] 1 1 0 0 1 0
head(ExampleDataBinary$ID)
#> [1] 1 1 1 1 1 1
head(ExampleDataBinary$Z)
#>          z1         z2          z3          z4          z5
#> 1 0.4222353  1.2901256  0.18618604  1.45383568  0.54869608
#> 2 1.5952225 -0.6379755  0.95796887 -0.03400311  0.09561614
#> 3 0.7988354 -0.7243116  1.00367469 -1.22336606 -0.16760091
#> 4 0.4850902  1.4883909  0.35125182  0.72302769 -0.83167772
#> 5 1.2920051  0.6865298 -0.88208441  0.21011013  0.40801256
#> 6 0.4446510 -0.5466993  0.02647889 -2.35711231  0.42658309