Example data with continuous outcomes

A simulated data set containing 7901 observations, 5 continuous covariates and 100 providers.

Usage

data(ExampleDataLinear)

Format

A list containing the following elements:

Y

a vector representing continuous outcomes. Generated as a linear combination of provider-specific effects, covariates, and a random error term (\(\epsilon\)), where \(\epsilon\) follows a normal distribution with mean 0 and standard deviation 1.

ID

a vector representing identifier for each provider. The number of individuals per provider is generated from a Poisson distribution with mean 80.

Z

a data frame containing 5 continuous variables. Generated from a multivariate normal distribution, where the mean is calculated as \((\gamma_i - \mu_{\gamma}) \cdot \rho / \sigma_{\gamma}\) with \(\mu_{\gamma} = \log(4/11)\) and \(\sigma_{\gamma} = 0.4\), and the correlation between covariates and provider effects is 0.1.

Examples

data(ExampleDataLinear)
head(ExampleDataLinear$Y)
#> [1]  2.341395  1.726385 -6.085487 -4.459568 -5.392451 -3.747529
head(ExampleDataLinear$ID)
#> [1] 1 1 1 1 1 1
head(ExampleDataLinear$Z)
#>           z1         z2          z3          z4          z5
#> 1  1.4371968  0.5030609  0.27799816  0.85035519  2.09705116
#> 2 -0.7367012  1.3282275  0.77115851  0.11971217  1.62687441
#> 3 -1.0977921  0.2010389 -1.33315545 -0.90291325 -1.17277380
#> 4 -0.1943289 -1.1456950 -1.55959090 -0.89686805 -1.00715112
#> 5 -1.3703641 -0.9746820 -0.05484035 -0.22093583 -0.73052043
#> 6 -0.4016549 -1.7874646 -1.54723763 -0.06955694 -0.01465253