Calculate direct/indirect standardized differences from a fitted linear_cre object

Provide direct/indirect standardized differences for a correlated random effect linear model.

Usage

# S3 method for class 'linear_cre'
SM_output(fit, parm, stdz = "indirect", ...)

Arguments

fit

a model fitted from linear_cre.

parm

specifies a subset of providers for which confidence intervals are to be given. By default, all providers are included. The class of parm should match the class of the provider IDs.

stdz

a character string or a vector specifying the standardization method(s). The possible values are:

"indirect" (default) indirect standardization method.

"direct" direct standardization method.

c("indirect", "direct") outputs both direct and indirect standardized measures.

...

additional arguments that can be passed to the function.

Value

A list containing the standardized differences based on the method(s) specified in stdz, as well as the observed and expected outcomes used to calculate the standardized measures:

indirect.difference

indirect standardized differences, if stdz includes "indirect".

direct.difference

direct standardized differences, if stdz includes "direct".

OE

a list of data frames containing the observed and expected outcomes used for calculating standardized measures.

Details

This function computes standardized differences for a random effect linear model using either direct or indirect methods, or both when specified. The function returns both the standardized differences and the observed and expected outcomes used for their calculation.

Examples

data(ExampleDataLinear)
outcome <- ExampleDataLinear$Y
covar <- ExampleDataLinear$Z
ProvID <- ExampleDataLinear$ProvID
data <- data.frame(outcome, ProvID, covar)
outcome.char <- colnames(data)[1]
ProvID.char <- colnames(data)[2]
wb.char <- c("z1", "z2")
other.char <- c("z3", "z4", "z5")
fit_cre <- linear_cre(data = data, Y.char = outcome.char, ProvID.char = ProvID.char,
wb.char = wb.char, other.char = other.char)
SM_output(fit_cre)
#> $indirect.difference
#>     Indirect_standardized.difference
#> 1                       6.785256e-03
#> 2                      -1.806919e-01
#> 3                      -3.780905e-01
#> 4                      -2.444168e-02
#> 5                       1.792029e-01
#> 6                      -5.833179e-02
#> 7                       1.404149e-01
#> 8                       1.708130e-01
#> 9                       1.438871e-02
#> 10                     -3.826240e-01
#> 11                     -2.380173e-01
#> 12                      1.130074e-01
#> 13                     -4.784690e-01
#> 14                     -2.121670e-01
#> 15                      1.316870e-01
#> 16                      4.548110e-01
#> 17                      1.404606e-01
#> 18                      2.728407e-02
#> 19                     -7.377754e-02
#> 20                     -2.436697e-01
#> 21                      2.623501e-01
#> 22                     -1.555494e-01
#> 23                     -9.184982e-02
#> 24                     -2.125215e-02
#> 25                      4.923645e-01
#> 26                     -3.477299e-01
#> 27                      1.984962e-01
#> 28                      1.776785e-01
#> 29                      1.442597e-01
#> 30                     -1.340119e-01
#> 31                     -5.967626e-02
#> 32                     -3.957903e-02
#> 33                     -3.879325e-01
#> 34                     -8.433291e-02
#> 35                      7.816037e-02
#> 36                     -2.720351e-01
#> 37                     -7.436285e-05
#> 38                     -1.710716e-01
#> 39                      1.310192e-01
#> 40                     -5.723146e-01
#> 41                      3.781653e-01
#> 42                     -2.360564e-01
#> 43                      2.745440e-03
#> 44                      1.065302e-01
#> 45                     -7.412840e-02
#> 46                     -9.871623e-02
#> 47                      1.104680e-01
#> 48                      3.032263e-01
#> 49                     -3.456566e-01
#> 50                      2.034062e-01
#> 51                      4.700280e-01
#> 52                     -2.538584e-01
#> 53                      1.319016e-01
#> 54                      3.282849e-01
#> 55                     -1.950201e-01
#> 56                      1.713150e-01
#> 57                     -1.142830e-01
#> 58                     -1.844562e-01
#> 59                      1.617155e-01
#> 60                      2.591615e-01
#> 61                     -8.460518e-02
#> 62                      2.233582e-01
#> 63                      3.969432e-01
#> 64                      1.989568e-02
#> 65                     -1.135930e-01
#> 66                      6.286360e-02
#> 67                     -3.409301e-02
#> 68                      1.529773e-01
#> 69                      1.300505e-02
#> 70                     -5.622530e-01
#> 71                      4.171539e-01
#> 72                      7.966611e-02
#> 73                      5.433189e-02
#> 74                      1.151334e-01
#> 75                      3.162371e-02
#> 76                     -2.810648e-01
#> 77                     -6.770220e-02
#> 78                     -2.146480e-02
#> 79                     -2.552357e-02
#> 80                     -1.155641e-01
#> 81                      1.100119e-01
#> 82                      8.167846e-02
#> 83                     -1.151083e-01
#> 84                      3.836972e-01
#> 85                     -1.218196e-01
#> 86                     -4.465184e-01
#> 87                      3.823881e-01
#> 88                     -8.371202e-02
#> 89                      4.387280e-01
#> 90                     -1.717491e-02
#> 91                     -2.121595e-01
#> 92                     -3.735372e-01
#> 93                     -1.436546e-01
#> 94                      1.707896e-01
#> 95                     -1.656838e-01
#> 96                     -2.082330e-01
#> 97                      2.012268e-01
#> 98                      4.135859e-01
#> 99                      1.216311e-01
#> 100                    -5.749045e-02
#> 
#> $OE
#> $OE$OE_indirect
#>              Obs         Exp
#> 1   -157.8165874 -158.420475
#> 2   -232.3446919 -217.527960
#> 3   -155.7116972 -126.976821
#> 4     12.5118082   14.320493
#> 5    -90.6949358 -105.568777
#> 6     49.2023397   53.868883
#> 7   -118.7168970 -129.107603
#> 8    -92.8830501 -104.839958
#> 9    -64.3941349  -65.516454
#> 10   -66.6317632  -36.787089
#> 11  -125.7666513 -108.391391
#> 12   -49.8459528  -58.434518
#> 13  -102.8353299  -64.557811
#> 14  -128.6105588 -115.031872
#> 15   -12.6446194  -22.916207
#> 16     3.3327061  -25.320384
#> 17  -100.9631388 -110.654921
#> 18  -181.4004275 -183.446732
#> 19    16.2163335   21.970982
#> 20  -111.5565978  -91.088346
#> 21     6.6193832  -11.745124
#> 22   -78.9118855  -66.467932
#> 23   -32.9822943  -27.103906
#> 24  -143.8097240 -142.130804
#> 25    31.1267540  -11.708959
#> 26  -105.6237387  -75.371235
#> 27     1.0776031  -13.611114
#> 28    14.3040052   -1.154020
#> 29   -92.3542080 -104.327767
#> 30  -110.7158070  -99.458811
#> 31   -44.8663247  -40.330929
#> 32  -149.0065605 -145.840238
#> 33   -10.1115421   21.698919
#> 34  -153.2465895 -146.584290
#> 35  -113.6828167 -119.544845
#> 36  -129.3169328 -106.193950
#> 37  -213.5935607 -213.587017
#> 38   -97.7134243  -85.225199
#> 39    17.3043342    5.643626
#> 40  -204.4398817 -162.088601
#> 41    40.7958557   16.971441
#> 42  -120.2125405 -106.285211
#> 43     6.6861449    6.463764
#> 44  -105.6510214 -115.132206
#> 45  -110.3801643 -103.560351
#> 46  -105.0063075  -97.602590
#> 47   -34.5007087  -43.448617
#> 48    55.2340545   27.640458
#> 49   -98.4461208  -68.719657
#> 50   -63.9994020  -78.848057
#> 51   -36.9757883  -80.218365
#> 52  -121.3701718 -103.346222
#> 53  -111.3923074 -121.548730
#> 54    31.7082314    3.147446
#> 55   -45.0191905  -27.467383
#> 56  -146.6153024 -158.607352
#> 57    52.7308077   61.530598
#> 58   -93.1716971  -74.910533
#> 59    -0.5295969  -12.334825
#> 60   -89.3370450 -105.923379
#> 61  -114.5095112 -107.910307
#> 62   -61.1226289  -79.438005
#> 63   -34.1180944  -61.507175
#> 64   -35.5741837  -37.046464
#> 65   -77.9057794  -69.499895
#> 66   -55.9309710  -60.142832
#> 67  -115.9937405 -113.470858
#> 68  -158.4234348 -170.508640
#> 69    -3.6874100   -4.688799
#> 70  -102.9104759  -62.428260
#> 71   -55.9813730  -92.273759
#> 72  -214.0818676 -220.853487
#> 73   -33.7430043  -37.872228
#> 74   -89.6277767 -100.680581
#> 75   -11.1482198  -14.089225
#> 76  -140.7043671 -119.343444
#> 77   -27.6742554  -24.018337
#> 78  -107.5128919 -105.516665
#> 79  -108.1703260 -106.205011
#> 80   -62.3823352  -53.368335
#> 81    67.6887596   58.337746
#> 82   -13.6584110  -20.192688
#> 83  -154.7615325 -146.358630
#> 84    16.2298715  -18.686577
#> 85   -19.4185380   -7.967498
#> 86  -139.4009886 -102.786476
#> 87   -68.1427562 -101.410519
#> 88   -83.0097163  -76.647603
#> 89    20.6483222  -21.469570
#> 90   -82.7261998  -81.300683
#> 91  -101.7475495  -87.108546
#> 92  -247.7483511 -219.359525
#> 93   100.3412472  111.402654
#> 94  -124.4328946 -139.633166
#> 95   -29.6111623  -16.522140
#> 96   -70.5064183  -52.181917
#> 97   -47.3160037  -59.590839
#> 98  -187.6258491 -225.262167
#> 99   -60.7712844  -69.163830
#> 100 -122.5673796 -117.738182
#> 
#>