Standardized Measures Output Function

Provide direct/indirect standardized difference for a fixed effect linear model.

Usage

# S3 method for linear_fe
SM_output(fit, parm, stdz = "indirect", null = "median")

Arguments

fit

a model fitted from linear_fe.

parm

specifies a subset of providers for which confidence intervals are to be given. By default, all providers are included. The class of parm should match the class of the provider IDs.

stdz

a character string or a vector specifying the standardization method(s). The possible values are:

"indirect" (default) indirect standardization method.

"direct" direct standardization method.

c("indirect", "direct") outputs both direct and indirect standardized measures.

null

a character string or a number defining the population norm. The default value is "median". The possible values are:

"median" the median of the provider effect estimates (\(\hat{\gamma}_i\)).

"mean" the weighted average of the provider effect estimates (\(\hat{\gamma}_i\)), where the weights correspond to the sample size of each provider.

numeric a user-defined numeric value representing the population norm.

Value

A list containing the standardized differences based on the method(s) specified in stdz, as well as the observed and expected outcomes used to calculate the standardized measures:

indirect.difference

indirect standardized differences, if stdz includes "indirect".

direct.difference

direct standardized differences, if stdz includes "direct".

OE

a list of data frames containing the observed and expected outcomes used for calculating standardized measures.

Details

This function computes standardized differences for a fixed effect linear model using either direct or indirect methods, or both when specified. For each method, the population norm is determined by the null argument. The population norm can be the median of the estimates, their weighted mean (with weights corresponding to provider sizes), or a user-defined numeric value.

Examples

data(ExampleDataLinear)
Y <- ExampleDataLinear$Y
Z <- ExampleDataLinear$Z
ID <- ExampleDataLinear$ID

# Fit a fixed linear effect model
fit_fe <- linear_fe(Y = Y, Z = Z, ID = ID)
#> Input format: Y, Z, and ID.
SM_output(fit_fe)
#> Error in SM_output(fit_fe): could not find function "SM_output"
SM_output(fit_fe, stdz = "direct", null = "mean")
#> Error in SM_output(fit_fe, stdz = "direct", null = "mean"): could not find function "SM_output"