Calculate direct/indirect standardized differences from a fitted linear_re object

Provide direct/indirect standardized differences for a random effect linear model.

Usage

# S3 method for linear_re
SM_output(fit, parm, stdz = "indirect")

Arguments

fit

a model fitted from linear_re.

parm

specifies a subset of providers for which confidence intervals are to be given. By default, all providers are included. The class of parm should match the class of the provider IDs.

stdz

a character string or a vector specifying the standardization method(s). The possible values are:

"indirect" (default) indirect standardization method.

"direct" direct standardization method.

c("indirect", "direct") outputs both direct and indirect standardized measures.

Value

A list containing the standardized differences based on the method(s) specified in stdz, as well as the observed and expected outcomes used to calculate the standardized measures:

indirect.difference

indirect standardized differences, if stdz includes "indirect".

direct.difference

direct standardized differences, if stdz includes "direct".

OE

a list of data frames containing the observed and expected outcomes used for calculating standardized measures.

Details

This function computes standardized differences for a random effect linear model using either direct or indirect methods, or both when specified. The function returns both the standardized differences and the observed and expected outcomes used for their calculation.

Examples

data(ExampleDataLinear)
outcome <- ExampleDataLinear$Y
covar <- ExampleDataLinear$Z
ID <- ExampleDataLinear$ID
fit_linear <- linear_re(Y = outcome, Z = covar, ID = ID)
#> Input format: Y, Z, and ID.
SM_output(fit_linear)
#> $indirect.difference
#>     Indirect_standardized.difference
#> 1                       -0.338240958
#> 2                       -0.794389223
#> 3                       -0.646058872
#> 4                        0.260222333
#> 5                        0.253447824
#> 6                        0.511178626
#> 7                       -0.061265301
#> 8                        0.006948695
#> 9                        0.038982628
#> 10                      -0.089775701
#> 11                      -0.403966192
#> 12                       0.258342952
#> 13                      -0.353891953
#> 14                      -0.480798587
#> 15                       0.228403421
#> 16                       0.652245404
#> 17                      -0.180256778
#> 18                      -0.508231543
#> 19                       0.288147391
#> 20                      -0.317763248
#> 21                       0.501834328
#> 22                      -0.054292981
#> 23                       0.161368104
#> 24                      -0.258343295
#> 25                       0.580695732
#> 26                      -0.290963078
#> 27                       0.524245442
#> 28                       0.475868844
#> 29                      -0.062803193
#> 30                      -0.224160405
#> 31                       0.074646030
#> 32                      -0.454060915
#> 33                       0.088741486
#> 34                      -0.369611317
#> 35                      -0.270597627
#> 36                      -0.324613417
#> 37                      -0.534115124
#> 38                      -0.285213112
#> 39                       0.354251227
#> 40                      -0.881389857
#> 41                       0.803002465
#> 42                      -0.517811990
#> 43                       0.293492361
#> 44                       0.035110130
#> 45                      -0.110060527
#> 46                      -0.211669804
#> 47                       0.241146710
#> 48                       0.621420984
#> 49                      -0.255612145
#> 50                       0.165658248
#> 51                       0.383515095
#> 52                      -0.392390843
#> 53                       0.068123077
#> 54                       0.679024992
#> 55                       0.048869088
#> 56                      -0.192475634
#> 57                       0.525940998
#> 58                      -0.197817532
#> 59                       0.448000079
#> 60                       0.020624157
#> 61                      -0.308840599
#> 62                       0.163885562
#> 63                       0.397151882
#> 64                       0.110659002
#> 65                      -0.215252603
#> 66                       0.110843986
#> 67                      -0.185974771
#> 68                      -0.341368033
#> 69                       0.320761519
#> 70                      -0.537901540
#> 71                       0.342325491
#> 72                      -0.528870103
#> 73                       0.240576715
#> 74                       0.159376193
#> 75                       0.315175243
#> 76                      -0.431149420
#> 77                       0.166323929
#> 78                       0.152830117
#> 79                      -0.047475405
#> 80                       0.005401941
#> 81                       0.711346301
#> 82                       0.115364610
#> 83                      -0.495816092
#> 84                       0.569053793
#> 85                       0.223492669
#> 86                      -0.543695052
#> 87                       0.393472922
#> 88                       0.008107068
#> 89                       0.633800387
#> 90                      -0.041475064
#> 91                      -0.310329631
#> 92                      -1.084831958
#> 93                       0.625010127
#> 94                      -0.119090926
#> 95                      -0.004338301
#> 96                      -0.123299039
#> 97                       0.287366666
#> 98                      -0.219943853
#> 99                       0.229334049
#> 100                     -0.272865484
#> 
#> $OE
#> $OE$OE_indirect
#>              Obs          Exp
#> 1   -155.6145089 -125.5110636
#> 2   -230.0948672 -164.9549509
#> 3   -157.3167533 -108.2162790
#> 4     10.5320226   -8.7244300
#> 5    -89.6064077 -110.6425771
#> 6     45.1599249    4.2656348
#> 7   -116.2496855 -111.7160532
#> 8    -90.3878779  -90.8742865
#> 9    -64.4224847  -67.4631298
#> 10   -71.7599760  -64.7574713
#> 11  -126.7941188  -97.3045868
#> 12   -49.7715221  -69.4055864
#> 13  -107.7470677  -79.4357115
#> 14  -128.7765779  -98.0054683
#> 15   -12.1066109  -29.9220778
#> 16     6.0394224  -35.0520380
#> 17   -97.7560404  -85.3183227
#> 18  -177.8377006 -139.7203349
#> 19    13.3299219   -9.1455746
#> 20  -113.2035787  -86.5114659
#> 21     7.4004609  -27.7279420
#> 22   -80.8861200  -76.5426815
#> 23   -35.3496778  -45.6772365
#> 24  -142.5206375 -122.1115172
#> 25    34.8360157  -15.6845130
#> 26  -108.9835809  -83.6697931
#> 27     0.7705653  -38.0235974
#> 28    13.9881378  -27.4124516
#> 29   -89.8203077  -84.6076428
#> 30  -111.3147221  -92.4852481
#> 31   -46.2168564  -51.8899546
#> 32  -146.7739652 -110.4490920
#> 33   -16.4276384  -23.7044403
#> 34  -152.2036768 -123.0043827
#> 35  -110.8402987  -90.5454766
#> 36  -131.3428784 -103.7507380
#> 37  -210.2765934 -163.2744625
#> 38   -98.4837831  -77.6632259
#> 39    17.0506167  -14.4777425
#> 40  -207.4692333 -142.2463839
#> 41    41.4270464   -9.1621089
#> 42  -120.5036274  -89.9527200
#> 43     4.9036865  -18.8691947
#> 44  -104.2860605 -107.4108621
#> 45  -110.7980111 -100.6724426
#> 46  -105.1583127  -89.2830774
#> 47   -34.3571870  -53.8900705
#> 48    55.8796613   -0.6696483
#> 49  -101.9947619  -80.0121174
#> 50   -62.0058309  -74.0988830
#> 51   -32.3735136  -67.6569024
#> 52  -122.7049541  -94.8452042
#> 53  -109.8554040 -115.1008809
#> 54    32.3683647  -26.7068096
#> 55   -48.2208443  -52.6190622
#> 56  -142.8737905 -129.4004961
#> 57    47.7652308    7.2677739
#> 58   -94.6838818  -75.0999462
#> 59    -0.9095513  -33.6135571
#> 60   -85.6139396  -86.9338856
#> 61  -113.8481691  -89.7586024
#> 62   -58.8215516  -72.2601677
#> 63   -30.6866006  -58.0900805
#> 64   -35.9659717  -44.1547379
#> 65   -78.2566016  -62.3279090
#> 66   -55.6853901  -63.1119372
#> 67  -115.3450471 -101.5829141
#> 68  -154.0294663 -127.0613917
#> 69    -5.4868106  -30.1854476
#> 70  -107.9241539  -69.1952431
#> 71   -51.9089448  -81.6912625
#> 72  -209.6125130 -164.6585542
#> 73   -34.4301435  -52.7139738
#> 74   -88.9069446 -104.2070591
#> 75   -12.6362373  -41.9475348
#> 76  -142.2026656 -109.4353097
#> 77   -29.7135840  -38.6950762
#> 78  -108.7812896 -122.9944905
#> 79  -108.2546845 -104.5990784
#> 80   -64.1332158  -64.5545672
#> 81    64.9044927    4.4400572
#> 82   -13.1613182  -22.3904870
#> 83  -153.3919327 -117.1973579
#> 84    18.3966230  -33.3872722
#> 85   -22.6172175  -43.6255284
#> 86  -142.6588204  -98.0758262
#> 87   -64.9047017  -99.1368459
#> 88   -84.3040691  -84.9202063
#> 89    23.2306267  -37.6142105
#> 90   -82.7237702  -79.2813398
#> 91  -102.9724576  -81.5597131
#> 92  -246.5599628 -164.1127340
#> 93    94.3237524   46.1979726
#> 94  -121.1550769 -110.5559844
#> 95   -32.0463300  -31.7036043
#> 96   -72.8348636  -61.9845482
#> 97   -46.1109061  -63.6402727
#> 98  -180.1134678 -160.0985772
#> 99   -60.3884917  -76.2125411
#> 100 -121.7265909  -98.8058902
#> 
#>