Calculate direct/indirect standardized ratios/rates from a fitted logis_re object

Provide direct/indirect standardized ratios/rates for a random effect logistic model.

Usage

# S3 method for class 'logis_re'
SM_output(
  fit,
  parm,
  stdz = "indirect",
  measure = c("rate", "ratio"),
  threads = 2,
  ...
)

Arguments

fit

a model fitted from logis_re.

parm

specifies a subset of providers for which confidence intervals are to be given. By default, all providers are included. The class of parm should match the class of the provider IDs.

stdz

a character string or a vector specifying the standardization method(s). The possible values are:

"indirect" (default) indirect standardization method.

"direct" direct standardization method.

c("indirect", "direct") outputs both direct and indirect standardized measures.

measure

a character string or a vector indicating whether the output measure is "ratio" or "rate"

"rate" output the standardized rate. The "rate" has been restricted to 0% - 100%.

"ratio" output the standardized ratio.

c("ratio", "rate") (default) output both the ratio and rate.

threads

an integer specifying the number of threads to use. The default value is 2.

...

additional arguments that can be passed to the function.

Value

A list contains standardized measures, as well as the observed and expected outcomes used for calculation, depending on the user's choice of standardization method (stdz) and measure type (measure).

indirect.ratio

standardization ratio using indirect method if stdz includes "indirect" and measure includes "ratio".

direct.ratio

standardization ratio using direct method if stdz includes "direct" and measure includes "ratio".

indirect.rate

standardization rate using indirect method if stdz includes "indirect" and measure includes "rate".

direct.rate

standardization rate using direct method if stdz includes "direct" and measure includes "rate".

OE

a list of data frames containing the observed and expected outcomes used for calculating standardized measures.

Examples

data(ExampleDataBinary)
outcome = ExampleDataBinary$Y
covar = ExampleDataBinary$Z
ProvID = ExampleDataBinary$ProvID
fit_re <- logis_re(Y = outcome, Z = covar, ProvID = ProvID)
#> Input format: Y, Z, and ProvID.
SR <- SM_output(fit_re, stdz = "direct", measure = "rate")
SR$direct.rate
#>     Direct_standardized.rate
#> 1                   35.94309
#> 2                   40.56492
#> 3                   36.16104
#> 4                   40.97619
#> 5                   38.46725
#> 6                   39.82165
#> 7                   39.37303
#> 8                   38.40258
#> 9                   31.02919
#> 10                  45.68533
#> 11                  43.09255
#> 12                  39.53139
#> 13                  41.81888
#> 14                  35.17798
#> 15                  39.49673
#> 16                  33.53254
#> 17                  43.41589
#> 18                  33.64967
#> 19                  40.37793
#> 20                  39.02882
#> 21                  43.67189
#> 22                  49.85904
#> 23                  39.93132
#> 24                  34.90648
#> 25                  34.37495
#> 26                  38.89604
#> 27                  41.25189
#> 28                  37.14199
#> 29                  40.60421
#> 30                  43.83157
#> 31                  39.97904
#> 32                  37.03225
#> 33                  37.40452
#> 34                  37.58911
#> 35                  43.13149
#> 36                  43.65465
#> 37                  35.48141
#> 38                  34.27404
#> 39                  36.79502
#> 40                  29.55985
#> 41                  31.55189
#> 42                  39.45095
#> 43                  42.75290
#> 44                  34.82813
#> 45                  44.10489
#> 46                  34.79794
#> 47                  41.57415
#> 48                  33.22127
#> 49                  32.87865
#> 50                  29.42144
#> 51                  34.48627
#> 52                  36.29749
#> 53                  40.12387
#> 54                  39.44504
#> 55                  35.22233
#> 56                  41.18194
#> 57                  41.16200
#> 58                  35.90947
#> 59                  39.77488
#> 60                  37.08551
#> 61                  47.95244
#> 62                  43.69869
#> 63                  41.36766
#> 64                  41.15840
#> 65                  34.65816
#> 66                  38.95322
#> 67                  36.05677
#> 68                  39.35616
#> 69                  36.50709
#> 70                  35.72051
#> 71                  37.43663
#> 72                  30.57373
#> 73                  34.86943
#> 74                  32.70136
#> 75                  45.91794
#> 76                  41.22473
#> 77                  34.58428
#> 78                  41.52656
#> 79                  38.56770
#> 80                  31.91444
#> 81                  35.19794
#> 82                  40.98767
#> 83                  42.33121
#> 84                  42.66968
#> 85                  43.20566
#> 86                  33.37243
#> 87                  37.32494
#> 88                  41.43719
#> 89                  36.29254
#> 90                  41.83808
#> 91                  41.37611
#> 92                  39.22900
#> 93                  34.94353
#> 94                  31.90025
#> 95                  42.80892
#> 96                  52.10416
#> 97                  41.53461
#> 98                  37.10010
#> 99                  37.78172
#> 100                 41.63882