Get confidence intervals for provider effects or standardized measures from a fitted linear_cre object

Provide confidence intervals for provider effects or standardized measures from a correlated random effect linear model.

Usage

# S3 method for class 'linear_cre'
confint(
  object,
  parm,
  level = 0.95,
  option = "SM",
  stdz = "indirect",
  alternative = "two.sided",
  ...
)

Arguments

object

a model fitted from linear_cre.

parm

specify a subset of providers for which confidence intervals are given. By default, all providers are included. The class of parm should match the class of the provider IDs.

level

the confidence level. The default value is 0.95.

option

a character string specifying whether the confidence intervals should be provided for provider effects or standardized measures:

"alpha" provider effect.

"SM" standardized measures.

stdz

a character string or a vector specifying the standardization method if option includes "SM". See stdz argument in SM_output.linear_re.

alternative

a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater", or "less". Note that "alpha" for argument option only supports "two.sided".

...

additional arguments that can be passed to the function.

Value

A list of data frames containing the confidence intervals based on the values of option and stdz.

CI.alpha

Confidence intervals for provider effects if option includes "alpha".

CI.indirect

Confidence intervals for indirect standardized differences if option includes "SM" and stdz includes "indirect".

CI.direct

Confidence intervals for direct standardized differences if option includes "SM" and stdz includes "direct".

Examples

data(ExampleDataLinear)
outcome <- ExampleDataLinear$Y
covar <- ExampleDataLinear$Z
ProvID <- ExampleDataLinear$ProvID
data <- data.frame(outcome, ProvID, covar)
outcome.char <- colnames(data)[1]
ProvID.char <- colnames(data)[2]
wb.char <- c("z1", "z2")
other.char <- c("z3", "z4", "z5")
fit_cre <- linear_cre(data = data, Y.char = outcome.char, ProvID.char = ProvID.char,
wb.char = wb.char, other.char = other.char)
confint(fit_cre)
#> $CI.indirect
#>     Indirect.Difference indirect.Lower indirect.Upper
#> 1          6.785256e-03   -0.185860942   0.1994314543
#> 2         -1.806919e-01   -0.380177809   0.0187940950
#> 3         -3.780905e-01   -0.584057702  -0.1721232381
#> 4         -2.444168e-02   -0.232713894   0.1838305265
#> 5          1.792029e-01   -0.019261190   0.3776670038
#> 6         -5.833179e-02   -0.259909718   0.1432461346
#> 7          1.404149e-01   -0.067857260   0.3486871602
#> 8          1.708130e-01   -0.042311306   0.3839372516
#> 9          1.438871e-02   -0.189348418   0.2181258351
#> 10        -3.826240e-01   -0.586361156  -0.1788869033
#> 11        -2.380173e-01   -0.447471372  -0.0285631664
#> 12         1.130074e-01   -0.092959798   0.3189746662
#> 13        -4.784690e-01   -0.680046911  -0.2768910584
#> 14        -2.121670e-01   -0.433250306   0.0089163439
#> 15         1.316870e-01   -0.072050104   0.3354241495
#> 16         4.548110e-01    0.232312419   0.6773094950
#> 17         1.404606e-01   -0.073930608   0.3548518375
#> 18         2.728407e-02   -0.179826037   0.2343941670
#> 19        -7.377754e-02   -0.277514666   0.1299595872
#> 20        -2.436697e-01   -0.441127455  -0.0462118843
#> 21         2.623501e-01    0.049225827   0.4754743847
#> 22        -1.555494e-01   -0.357127339   0.0460285137
#> 23        -9.184982e-02   -0.312933141   0.1292335092
#> 24        -2.125215e-02   -0.223901049   0.1813967509
#> 25         4.923645e-01    0.297836236   0.6868928000
#> 26        -3.477299e-01   -0.542258210  -0.1532016455
#> 27         1.984962e-01   -0.009776031   0.4067683892
#> 28         1.776785e-01   -0.016849831   0.3722067329
#> 29         1.442597e-01   -0.054204355   0.3427238379
#> 30        -1.340119e-01   -0.331469640   0.0634459306
#> 31        -5.967626e-02   -0.265643487   0.1462909766
#> 32        -3.957903e-02   -0.241156955   0.1619988976
#> 33        -3.879325e-01   -0.587418402  -0.1884464983
#> 34        -8.433291e-02   -0.286981809   0.1183159900
#> 35         7.816037e-02   -0.128949730   0.2852704737
#> 36        -2.720351e-01   -0.468501717  -0.0755684610
#> 37        -7.436285e-05   -0.193654741   0.1935060156
#> 38        -1.710716e-01   -0.380525676   0.0383825296
#> 39         1.310192e-01   -0.061627010   0.3236653858
#> 40        -5.723146e-01   -0.780586816  -0.3640423957
#> 41         3.781653e-01    0.155666769   0.6006638454
#> 42        -2.360564e-01   -0.464500374  -0.0076124773
#> 43         2.745440e-03   -0.197778315   0.2032691954
#> 44         1.065302e-01   -0.086116034   0.2991763621
#> 45        -7.412840e-02   -0.264051175   0.1157943657
#> 46        -9.871623e-02   -0.305826336   0.1083938675
#> 47         1.104680e-01   -0.090055756   0.3109917546
#> 48         3.032263e-01    0.112408631   0.4940440427
#> 49        -3.456566e-01   -0.541146809  -0.1501663122
#> 50         2.034062e-01   -0.006047873   0.4128603323
#> 51         4.700280e-01    0.280105234   0.6599507751
#> 52        -2.538584e-01   -0.465737983  -0.0419789150
#> 53         1.319016e-01   -0.072941480   0.3367446707
#> 54         3.282849e-01    0.133756609   0.5228131730
#> 55        -1.950201e-01   -0.386745497  -0.0032946711
#> 56         1.713150e-01   -0.041809291   0.3844392669
#> 57        -1.142830e-01   -0.319126071   0.0905600792
#> 58        -1.844562e-01   -0.368448261  -0.0004641474
#> 59         1.617155e-01   -0.047738651   0.3711695545
#> 60         2.591615e-01    0.038078149   0.4802447994
#> 61        -8.460518e-02   -0.288342306   0.1191319474
#> 62         2.233582e-01    0.023872291   0.4228441947
#> 63         3.969432e-01    0.182551971   0.6113344161
#> 64         1.989568e-02   -0.188376532   0.2281678885
#> 65        -1.135930e-01   -0.321865243   0.0946791770
#> 66         6.286360e-02   -0.154130803   0.2798580095
#> 67        -3.409301e-02   -0.242365222   0.1741791986
#> 68         1.529773e-01   -0.049671616   0.3556261832
#> 69         1.300505e-02   -0.191838023   0.2178481275
#> 70        -5.622530e-01   -0.772909346  -0.3515966519
#> 71         4.171539e-01    0.222625586   0.6116821497
#> 72         7.966611e-02   -0.116800518   0.2761327386
#> 73         5.433189e-02   -0.151635341   0.2602991232
#> 74         1.151334e-01   -0.071331570   0.3015983308
#> 75         3.162371e-02   -0.157416599   0.2206640206
#> 76        -2.810648e-01   -0.487032010  -0.0750975456
#> 77        -6.770220e-02   -0.304294564   0.1688901680
#> 78        -2.146480e-02   -0.210505114   0.1675755052
#> 79        -2.552357e-02   -0.230366649   0.1793195018
#> 80        -1.155641e-01   -0.319301232   0.0881730217
#> 81         1.100119e-01   -0.086454701   0.3064785556
#> 82         8.167846e-02   -0.119899469   0.2832563835
#> 83        -1.151083e-01   -0.324562360   0.0943458452
#> 84         3.836972e-01    0.192879527   0.5745149386
#> 85        -1.218196e-01   -0.309989608   0.0663504654
#> 86        -4.465184e-01   -0.646004394  -0.2470324903
#> 87         3.823881e-01    0.187859797   0.5769163617
#> 88        -8.371202e-02   -0.289679253   0.1222552109
#> 89         4.387280e-01    0.252263098   0.6251929986
#> 90        -1.717491e-02   -0.215639003   0.1812891907
#> 91        -2.121595e-01   -0.426550687   0.0022317584
#> 92        -3.735372e-01   -0.579504419  -0.1675699550
#> 93        -1.436546e-01   -0.348497715   0.0611884354
#> 94         1.707896e-01   -0.021856629   0.3634357668
#> 95        -1.656838e-01   -0.368332721   0.0369650787
#> 96        -2.082330e-01   -0.401813344  -0.0146525873
#> 97         2.012268e-01   -0.024185643   0.4266392759
#> 98         4.135859e-01    0.222768208   0.6044036201
#> 99         1.216311e-01   -0.092760129   0.3360223161
#> 100       -5.749045e-02   -0.254948235   0.1399673360
#>