Get confidence intervals for provider effects or standardized measures from a fitted logis_cre object

Provide confidence intervals for provider effects or standardized measures from a correlated random effect logistic model.

Usage

# S3 method for class 'logis_cre'
confint(
  object,
  parm,
  level = 0.95,
  option = "SM",
  measure = c("rate", "ratio"),
  stdz = "indirect",
  alternative = "two.sided",
  ...
)

Arguments

object

a model fitted from logis_cre.

parm

specify a subset of providers for which confidence intervals are given. By default, all providers are included. The class of parm should match the class of the provider IDs.

level

the confidence level. The default value is 0.95.

option

a character string specifying whether the confidence intervals should be provided for provider effects or standardized measures:

"alpha" provider effect.

"SM" standardized measures.

measure

a character string or a vector indicating whether the output measure is "ratio" or "rate" if option = "SM". Both "rate" and "ratio" will be provided by default.

stdz

a character string or a vector specifying the standardization method if option includes "SM". See stdz argument in SM_output.linear_re.

alternative

a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater", or "less". Note that "alpha" for argument option only supports "two.sided".

...

additional arguments that can be passed to the function.

Value

A list of data frames containing the confidence intervals based on the values of option and stdz.

CI.alpha

Confidence intervals for provider effects if option includes "alpha".

CI.indirect

Confidence intervals for indirect standardized differences if option includes "SM" and stdz includes "indirect".

CI.direct

Confidence intervals for direct standardized differences if option includes "SM" and stdz includes "direct".

Examples

data(ExampleDataBinary)
outcome <- ExampleDataBinary$Y
covar <- ExampleDataBinary$Z
ProvID <- ExampleDataBinary$ProvID
data <- data.frame(outcome, ProvID, covar)
outcome.char <- colnames(data)[1]
ProvID.char <- colnames(data)[2]
wb.char <- c("z1", "z2")
other.char <- c("z3", "z4", "z5")
fit_cre <- logis_cre(data = data, Y.char = outcome.char, ProvID.char = ProvID.char,
wb.char = wb.char, other.char = other.char)
confint(fit_cre)
#> $CI.indirect_ratio
#>     Indirect.Ratio Ratio.Lower Ratio.Upper
#> 1        0.9732355   0.8905384    1.056724
#> 2        1.0534595   0.8871499    1.230996
#> 3        0.9982964   0.8957251    1.104824
#> 4        0.9806159   0.8906880    1.072319
#> 5        1.0035894   0.8774272    1.133516
#> 6        1.0070578   0.9104319    1.103071
#> 7        1.0617226   0.9190970    1.215509
#> 8        1.0336831   0.9086970    1.163638
#> 9        0.9095868   0.7653760    1.066216
#> 10       1.0095990   0.9340317    1.085327
#> 11       1.0468042   0.9244687    1.173829
#> 12       0.9842014   0.8882924    1.081309
#> 13       0.9852279   0.9172799    1.050782
#> 14       0.9683631   0.8586582    1.083701
#> 15       1.0036205   0.8950931    1.116136
#> 16       0.9705700   0.8308825    1.120012
#> 17       1.0585839   0.9440679    1.180756
#> 18       0.9656561   0.8438096    1.090817
#> 19       1.0218657   0.9157870    1.130303
#> 20       0.9814979   0.8929212    1.069473
#> 21       1.0119664   0.9383833    1.084375
#> 22       1.0006846   0.9922632    1.007020
#> 23       0.9994158   0.9236468    1.075439
#> 24       0.9574055   0.8547773    1.064656
#> 25       0.9524799   0.8397130    1.076165
#> 26       1.0001726   0.9035029    1.099245
#> 27       1.0394683   0.9323075    1.148264
#> 28       0.9799667   0.8725331    1.090383
#> 29       0.9799967   0.8934694    1.068670
#> 30       1.0650262   0.9679020    1.163873
#> 31       1.0065321   0.9302855    1.084596
#> 32       0.9599395   0.8691282    1.053149
#> 33       0.9988531   0.9077969    1.094193
#> 34       0.9966134   0.8972986    1.098988
#> 35       1.0481424   0.9394784    1.160449
#> 36       0.9994847   0.9418749    1.053815
#> 37       1.0242697   0.8667671    1.194977
#> 38       1.0488527   0.8597137    1.267326
#> 39       0.9902120   0.8804785    1.106410
#> 40       0.9658378   0.8157622    1.153189
#> 41       0.9207539   0.8004900    1.045345
#> 42       1.0373346   0.9305328    1.148487
#> 43       1.0156192   0.9309804    1.100178
#> 44       1.0334423   0.8650413    1.215863
#> 45       1.0303961   0.9344517    1.125920
#> 46       0.9537837   0.8493693    1.061390
#> 47       1.0425093   0.9363767    1.152437
#> 48       0.9527351   0.8216093    1.093551
#> 49       0.9825964   0.7148860    1.317346
#> 50       0.9164852   0.7837396    1.062481
#> 51       1.0146538   0.8592439    1.181800
#> 52       0.9623261   0.8489713    1.079976
#> 53       1.0547837   0.9289540    1.185589
#> 54       0.9936394   0.9129070    1.076329
#> 55       0.9700595   0.8641254    1.080484
#> 56       1.0597924   0.9388966    1.185525
#> 57       1.0576708   0.9470071    1.177219
#> 58       0.9608336   0.8577013    1.067647
#> 59       0.9951810   0.9017969    1.090847
#> 60       0.9770425   0.8782038    1.079382
#> 61       0.9992555   0.9485888    1.045360
#> 62       1.0583850   0.9647954    1.150507
#> 63       1.0541551   0.9341418    1.179925
#> 64       1.0380975   0.9181385    1.163295
#> 65       0.9858475   0.8551710    1.121748
#> 66       0.9919405   0.8954827    1.087470
#> 67       1.0034756   0.8569679    1.164761
#> 68       0.9604757   0.8869711    1.034153
#> 69       0.9669722   0.8592114    1.078362
#> 70       0.9959621   0.8755965    1.124415
#> 71       1.0155983   0.8846222    1.153897
#> 72       0.9180130   0.7538213    1.103029
#> 73       0.9592810   0.8469444    1.075807
#> 74       0.9786539   0.8300390    1.142340
#> 75       1.0674667   0.9686174    1.165427
#> 76       1.0312300   0.9211962    1.144160
#> 77       0.9897164   0.8484697    1.144350
#> 78       0.9937059   0.9139253    1.072905
#> 79       1.0359878   0.9199775    1.157282
#> 80       0.9072836   0.7911514    1.028465
#> 81       0.9926156   0.6640434    1.478806
#> 82       1.0607208   0.9384973    1.189800
#> 83       0.9940930   0.9231778    1.061687
#> 84       1.0069340   0.9221931    1.091485
#> 85       1.0682034   0.9642014    1.174567
#> 86       0.9480966   0.8301904    1.073885
#> 87       0.9723217   0.8647236    1.082881
#> 88       1.0191684   0.9233556    1.113777
#> 89       0.9921391   0.8802933    1.108959
#> 90       1.0037484   0.8990812    1.111324
#> 91       0.9819312   0.9085115    1.054590
#> 92       0.9813863   0.8951208    1.066633
#> 93       1.0300894   0.8783728    1.201959
#> 94       0.9462932   0.8456094    1.048098
#> 95       1.0162847   0.9328425    1.098521
#> 96       1.0012724   0.9910684    1.008962
#> 97       1.0144616   0.9247218    1.104538
#> 98       1.0080771   0.8896430    1.136280
#> 99       1.0065332   0.8958970    1.122022
#> 100      1.0304649   0.9273641    1.135923
#> 
#> $CI.indirect_rate
#>     Indirect.Rate Rate.Lower Rate.Upper
#> 1        37.73370   34.52742   40.97068
#> 2        40.84410   34.39605   47.72742
#> 3        38.70535   34.72851   42.83556
#> 4        38.01985   34.53322   41.57530
#> 5        38.91057   34.01908   43.94801
#> 6        39.04504   35.29872   42.76760
#> 7        41.16447   35.63468   47.12700
#> 8        40.07734   35.23145   45.11588
#> 9        35.26596   29.67470   41.33867
#> 10       39.14357   36.21372   42.07964
#> 11       40.58606   35.84294   45.51099
#> 12       38.15887   34.44034   41.92386
#> 13       38.19867   35.56423   40.74029
#> 14       37.54479   33.29138   42.01661
#> 15       38.91177   34.70401   43.27416
#> 16       37.63036   32.21448   43.42445
#> 17       41.04278   36.60283   45.77958
#> 18       37.43984   32.71568   42.29251
#> 19       39.61916   35.50634   43.82341
#> 20       38.05405   34.61981   41.46498
#> 21       39.23535   36.38244   42.04274
#> 22       38.79794   38.47143   39.04359
#> 23       38.74875   35.81108   41.69629
#> 24       37.11995   33.14091   41.27819
#> 25       36.92898   32.55685   41.72441
#> 26       38.77809   35.03007   42.61928
#> 27       40.30164   36.14687   44.51980
#> 28       37.99468   33.82933   42.27569
#> 29       37.99585   34.64106   41.43382
#> 30       41.29256   37.52692   45.12500
#> 31       39.02466   36.06847   42.05132
#> 32       37.21820   33.69732   40.83205
#> 33       38.72693   35.19655   42.42340
#> 34       38.64010   34.78952   42.60929
#> 35       40.63795   36.42489   44.99221
#> 36       38.75142   36.51781   40.85788
#> 37       39.71237   33.60577   46.33093
#> 38       40.66549   33.33231   49.13602
#> 39       38.39190   34.13738   42.89706
#> 40       37.44688   31.62824   44.71074
#> 41       35.69892   31.03612   40.52949
#> 42       40.21892   36.07806   44.52843
#> 43       39.37698   36.09541   42.65545
#> 44       40.06800   33.53886   47.14072
#> 45       39.94990   36.23000   43.65349
#> 46       36.97953   32.93124   41.15159
#> 47       40.41955   36.30464   44.68158
#> 48       36.93888   31.85494   42.39851
#> 49       38.09664   27.71713   51.07537
#> 50       35.53341   30.38668   41.19388
#> 51       39.33955   33.31409   45.82004
#> 52       37.31073   32.91581   41.87218
#> 53       40.89544   36.01685   45.96694
#> 54       38.52479   35.39468   41.73077
#> 55       37.61057   33.50335   41.89188
#> 56       41.08964   36.40234   45.96445
#> 57       41.00738   36.71679   45.64245
#> 58       37.25286   33.25428   41.39416
#> 59       38.58456   34.96393   42.29365
#> 60       37.88131   34.04919   41.84914
#> 61       38.74253   36.77812   40.53006
#> 62       41.03507   37.40647   44.60678
#> 63       40.87107   36.21798   45.74734
#> 64       40.24849   35.59751   45.10258
#> 65       38.22269   33.15618   43.49175
#> 66       38.45892   34.71912   42.16275
#> 67       38.90615   33.22584   45.15940
#> 68       37.23899   34.38911   40.09556
#> 69       37.49086   33.31283   41.80959
#> 70       38.61484   33.94810   43.59513
#> 71       39.37617   34.29804   44.73820
#> 72       35.59265   29.22671   42.76596
#> 73       37.19267   32.83722   41.71055
#> 74       37.94378   32.18177   44.29014
#> 75       41.38718   37.55465   45.18522
#> 76       39.98223   35.71607   44.36068
#> 77       38.37269   32.89636   44.36804
#> 78       38.52737   35.43416   41.59804
#> 79       40.16670   35.66882   44.86944
#> 80       35.17665   30.67405   39.87503
#> 81       38.48510   25.74589   57.33539
#> 82       41.12563   36.38685   46.13021
#> 83       38.54238   35.79290   41.16311
#> 84       39.04024   35.75472   42.31839
#> 85       41.41574   37.38344   45.53959
#> 86       36.75903   32.18764   41.63603
#> 87       37.69827   33.52655   41.98483
#> 88       39.51458   35.79979   43.18268
#> 89       38.46662   34.13021   42.99588
#> 90       38.91673   34.85863   43.08758
#> 91       38.07085   35.22426   40.88794
#> 92       38.04972   34.70509   41.35484
#> 93       39.93801   34.05574   46.60164
#> 94       36.68911   32.78546   40.63622
#> 95       39.40278   36.16761   42.59120
#> 96       38.82073   38.42511   39.11885
#> 97       39.33210   35.85276   42.82450
#> 98       39.08456   34.49271   44.05515
#> 99       39.02470   34.73518   43.50235
#> 100      39.95257   35.95520   44.04132
#>