Get confidence intervals for provider effects or standardized measures from a fitted logis_fe object

Provide confidence intervals for provider effects or standardized measures from a fixed effect logistic model.

Usage

# S3 method for class 'logis_fe'
confint(
  object,
  parm,
  level = 0.95,
  test = "exact",
  option = "SM",
  stdz = "indirect",
  null = "median",
  measure = c("rate", "ratio"),
  alternative = "two.sided",
  ...
)

Arguments

object

a model fitted from logis_fe.

parm

specify a subset of providers for which confidence intervals are given. By default, all providers are included. The class of parm should match the class of the provider IDs.

level

the confidence level. The default value is 0.95.

test

a character string specifying the type of testing method. The default is "exact".

"exact" exact test.

"wald" wald test.

"score" score test.

option

a character string specifying whether the confidence intervals should be provided for provider effects or standardized measures:

"gamma" provider effect.

"SM" standardized measures.

stdz

a character string or a vector specifying the standardization method if option = "SM". See stdz argument in SM_output.logis_fe.

null

a character string or a number defining the population norm if option = "SM".

measure

a character string or a vector indicating whether the output measure is "ratio" or "rate" if option = "SM". Both "rate" and "ratio" will be provided by default.

"rate" output the standardized rate. The "rate" has been restricted to 0% - 100%.

"ratio" output the standardized ratio.

c("ratio", "rate") output both the standardized rate and ratio.

alternative

a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater", or "less". Note that "gamma" for argument option only supports "two.sided".

...

additional arguments that can be passed to the function.

Value

A dataframe (option = "gamma") or a list of data frames (option = "SM") containing the point estimate, and lower and upper bounds of the estimate.

Details

The wald test is invalid for extreme providers (i.e. when provider effect goes to infinity). We suggest using score or exact test to generate confidence intervals.

Examples

data(ExampleDataBinary)
outcome = ExampleDataBinary$Y
covar = ExampleDataBinary$Z
ProvID = ExampleDataBinary$ProvID
fit_fe <- logis_fe(Y = outcome, Z = covar, ProvID = ProvID, message = FALSE)
confint(fit_fe, option = "gamma")
#>            gamma gamma.lower  gamma.upper
#> 1    -1.34797989  -2.0537002 -0.646004794
#> 2    -0.67625337  -1.4783625  0.108280207
#> 3    -1.28525031  -1.8934880 -0.687370098
#> 4    -0.67938233  -1.2797033 -0.086092562
#> 5    -0.99265740  -1.6883067 -0.306586937
#> 6    -0.81532974  -1.4691758 -0.158365019
#> 7    -0.87543592  -1.5047498 -0.261985376
#> 8    -1.00068080  -1.6532438 -0.357317086
#> 9    -2.01487404  -2.7118054 -1.356036731
#> 10   -0.03521049  -0.7151699  0.646111807
#> 11   -0.41361385  -1.0331518  0.199699393
#> 12   -0.85510502  -1.4958132 -0.220165172
#> 13   -0.53422603  -1.2356102  0.178346095
#> 14   -1.44522997  -2.1448296 -0.766708116
#> 15   -0.86171651  -1.4996089 -0.234152168
#> 16   -1.67040261  -2.3682618 -0.997904352
#> 17   -0.38402254  -0.9825887  0.203794375
#> 18   -1.68115574  -2.3994015 -0.975923171
#> 19   -0.76531946  -1.3164409 -0.218223501
#> 20   -0.92383165  -1.5190103 -0.328029770
#> 21   -0.31598883  -0.9757455  0.350997995
#> 22    2.10966982   0.3349658  5.209107007
#> 23   -0.79288236  -1.4900974 -0.096824052
#> 24   -1.45276650  -2.0803594 -0.839562234
#> 25   -1.52753761  -2.1798670 -0.905815116
#> 26   -0.93347260  -1.6273596 -0.247829872
#> 27   -0.61772976  -1.3071059  0.071794357
#> 28   -1.16019294  -1.7779386 -0.551318825
#> 29   -0.71185643  -1.3658122 -0.066162358
#> 30   -0.33318626  -0.9325393  0.265798379
#> 31   -0.77986366  -1.5054646 -0.062132886
#> 32   -1.17567914  -1.8019958 -0.558866645
#> 33   -1.14317960  -1.8612181 -0.441887844
#> 34   -1.10685453  -1.7255004 -0.497054091
#> 35   -0.36491386  -1.0532923  0.313468992
#> 36   -0.26419627  -0.9965540  0.491294276
#> 37   -1.41172803  -2.1378951 -0.712369181
#> 38   -1.58135752  -2.3192928 -0.889554893
#> 39   -1.21308490  -1.8726683 -0.572213081
#> 40  -10.71003595        -Inf -4.449114094
#> 41   -2.04040784  -2.8338347 -1.284182751
#> 42   -0.86612613  -1.5067837 -0.235425596
#> 43   -0.42229711  -1.0979568  0.254360346
#> 44   -1.49821614  -2.2055709 -0.815138939
#> 45   -0.28407706  -0.9057786  0.341775719
#> 46   -1.49373950  -2.1736451 -0.827343803
#> 47   -0.59024463  -1.2395408  0.051745937
#> 48   -1.70754053  -2.3941624 -1.048207588
#> 49   -8.08483094        -Inf -2.327590188
#> 50   -2.39811515  -3.2284479 -1.618721988
#> 51   -1.55619892  -2.2789772 -0.858229080
#> 52   -1.29496181  -2.0038880 -0.603276054
#> 53   -0.77298706  -1.4471202 -0.103876054
#> 54   -0.85328390  -1.6009438 -0.116481623
#> 55   -1.41767677  -2.0574086 -0.792812684
#> 56   -0.66242978  -1.2398346 -0.091294432
#> 57   -0.64115255  -1.2957887 -0.006435919
#> 58   -1.32646278  -1.9666843 -0.699596625
#> 59   -0.82199750  -1.4738207 -0.178134425
#> 60   -1.17485306  -1.8346318 -0.526685839
#> 61    0.22681516  -0.4295697  0.913826264
#> 62   -0.30464309  -0.9726764  0.373924824
#> 63   -0.58766599  -1.3127715  0.122904080
#> 64   -0.63497696  -1.3084996  0.028114194
#> 65   -1.56981768  -2.3580068 -0.800869968
#> 66   -0.93148088  -1.5457470 -0.315060646
#> 67   -1.37293507  -2.2276349 -0.564528600
#> 68   -0.87563371  -1.5238313 -0.231888314
#> 69   -1.24453551  -1.8737814 -0.628077460
#> 70   -1.37964221  -2.1055889 -0.677810694
#> 71   -1.13189518  -1.8252968 -0.455360436
#> 72   -2.27495244  -3.1825218 -1.444903513
#> 73   -1.45420261  -2.0742817 -0.847440975
#> 74   -1.83001701  -2.5858310 -1.112767866
#> 75   -0.11220514  -0.6757168  0.460139170
#> 76   -0.63525849  -1.2852645  0.009311440
#> 77   -1.55579587  -2.3180779 -0.830171360
#> 78   -0.60979649  -1.2215577  0.003150925
#> 79   -0.97870858  -1.6492041 -0.321498826
#> 80   -1.95822469  -2.7208826 -1.228884058
#> 81   -6.23331119        -Inf -0.768916759
#> 82   -0.67639817  -1.2911662 -0.072201714
#> 83   -0.51406360  -1.1155439  0.100016951
#> 84   -0.46296110  -1.0851628  0.160617546
#> 85   -0.39609871  -1.0184524  0.224715876
#> 86   -1.75831645  -2.5468001 -0.999400891
#> 87   -1.13991691  -1.7742716 -0.516148679
#> 88   -0.63843087  -1.1958845 -0.076366251
#> 89   -1.28315787  -1.9520393 -0.628462126
#> 90   -0.55396407  -1.2064798  0.090692052
#> 91   -0.61513058  -1.2592403  0.031179117
#> 92   -0.89304494  -1.5437139 -0.239978505
#> 93   -1.48728104  -2.2268665 -0.791843698
#> 94   -2.04000365  -2.8684329 -1.230668884
#> 95   -0.46827940  -1.0447749  0.112325167
#> 96    2.43108467   0.6994034  5.498454728
#> 97   -0.61918928  -1.2021864 -0.036281584
#> 98   -1.19743934  -1.9854846 -0.441504316
#> 99   -1.07806889  -1.6874791 -0.480059312
#> 100  -0.59084370  -1.2185063  0.033368998
confint(fit_fe, option = "SM")
#> $CI.indirect_ratio
#>     indirect_ratio CI_ratio.lower CI_ratio.upper
#> 1        0.9025369      0.7417186      1.0659296
#> 2        1.1081224      0.7780800      1.4599834
#> 3        0.8994107      0.7377677      1.0693919
#> 4        1.0742682      0.8982112      1.2564651
#> 5        0.9771412      0.7431920      1.2206725
#> 6        1.0311867      0.8501812      1.2097002
#> 7        1.0199439      0.7928109      1.2647141
#> 8        0.9761538      0.7710886      1.1899710
#> 9        0.5975821      0.4004864      0.8296783
#> 10       1.2230053      1.0520744      1.3928684
#> 11       1.1930386      0.9618795      1.4311229
#> 12       1.0216143      0.8357312      1.2124946
#> 13       1.0876559      0.9299446      1.2384005
#> 14       0.8432444      0.6531977      1.0508678
#> 15       1.0211703      0.8233102      1.2287191
#> 16       0.7402283      0.5302923      0.9740582
#> 17       1.1920803      0.9817006      1.4139520
#> 18       0.7685535      0.5634987      0.9848703
#> 19       1.0523600      0.8774389      1.2310994
#> 20       1.0010548      0.8375043      1.1641068
#> 21       1.1398736      0.9889740      1.2860926
#> 22       1.1099878      1.0677742      1.1223934
#> 23       1.0295094      0.8777374      1.1820287
#> 24       0.8443623      0.6736492      1.0271058
#> 25       0.8047108      0.6281825      1.0077050
#> 26       0.9984046      0.8127460      1.1906228
#> 27       1.0900073      0.8913082      1.2886455
#> 28       0.9269613      0.7414760      1.1205157
#> 29       1.0600648      0.8814338      1.2454315
#> 30       1.1666002      0.9986496      1.3348419
#> 31       1.0318668      0.8790236      1.1895100
#> 32       0.9305034      0.7628229      1.1058468
#> 33       0.9462238      0.7781190      1.1261303
#> 34       0.9491952      0.7811795      1.1243982
#> 35       1.1746635      0.9622184      1.3962924
#> 36       1.1211651      0.9867277      1.2441023
#> 37       0.8208371      0.5846392      1.0841298
#> 38       0.7279144      0.4876113      1.0173623
#> 39       0.9117180      0.7249289      1.1147446
#> 40       0.0000000      0.0000000      0.2191535
#> 41       0.6493380      0.4459617      0.8824974
#> 42       1.0177836      0.8384266      1.2055589
#> 43       1.1288054      0.9566697      1.3002308
#> 44       0.7866672      0.5540213      1.0442227
#> 45       1.1944312      1.0065906      1.3786573
#> 46       0.8320372      0.6446986      1.0306189
#> 47       1.1013911      0.9093049      1.2999171
#> 48       0.7314334      0.5319425      0.9559344
#> 49       0.0000000      0.0000000      0.5179884
#> 50       0.5399413      0.3547845      0.7618012
#> 51       0.7767765      0.5516960      1.0259254
#> 52       0.8794900      0.6647526      1.1101340
#> 53       1.0503812      0.8357049      1.2698771
#> 54       1.0174328      0.8483534      1.1937618
#> 55       0.8540790      0.6791876      1.0414258
#> 56       1.0916279      0.8955911      1.2940340
#> 57       1.0894223      0.8928405      1.3044097
#> 58       0.8792143      0.6987685      1.0711812
#> 59       1.0297793      0.8511476      1.2149037
#> 60       0.9290187      0.7488682      1.1187555
#> 61       1.2215167      1.1015182      1.3265864
#> 62       1.1556286      0.9885338      1.3153425
#> 63       1.1084466      0.8826186      1.3477723
#> 64       1.0995326      0.8757324      1.3337620
#> 65       0.7971367      0.5698776      1.0411913
#> 66       0.9989035      0.8227287      1.1733580
#> 67       0.8383100      0.5764216      1.1452386
#> 68       1.0126633      0.8575288      1.1709319
#> 69       0.8989515      0.7114120      1.0980551
#> 70       0.8583735      0.6549620      1.0825775
#> 71       0.9299052      0.7079551      1.1684102
#> 72       0.4980039      0.2777299      0.7815725
#> 73       0.8321012      0.6498947      1.0262905
#> 74       0.6840262      0.4731590      0.9292337
#> 75       1.2535229      1.0787469      1.4205431
#> 76       1.0925784      0.8897916      1.3003877
#> 77       0.7787017      0.5543629      1.0370969
#> 78       1.0815606      0.9254043      1.2368520
#> 79       0.9845495      0.7904178      1.1902884
#> 80       0.6651076      0.4614515      0.8967515
#> 81       0.0000000      0.0000000      1.1422604
#> 82       1.0857752      0.8814482      1.3007155
#> 83       1.0965577      0.9548462      1.2299472
#> 84       1.1255589      0.9578273      1.2922141
#> 85       1.1540167      0.9741097      1.3352526
#> 86       0.7420399      0.5322989      0.9761918
#> 87       0.9317680      0.7381823      1.1351108
#> 88       1.0855981      0.9203575      1.2468592
#> 89       0.8925744      0.7035890      1.0938009
#> 90       1.1243234      0.9107326      1.3456480
#> 91       1.0754348      0.9198150      1.2295322
#> 92       1.0092077      0.8359519      1.1806806
#> 93       0.7932752      0.5735091      1.0555737
#> 94       0.7136044      0.5197124      0.9209132
#> 95       1.1211595      0.9688390      1.2703068
#> 96       1.1416116      1.0977767      1.1540773
#> 97       1.0853426      0.9246933      1.2460695
#> 98       0.9157454      0.6972095      1.1617964
#> 99       0.9530179      0.7718654      1.1444547
#> 100      1.1022579      0.9138391      1.2936799
#> 
#> $CI.indirect_rate
#>     indirect_rate CI_rate.lower CI_rate.upper
#> 1        34.99262      28.75747     41.327582
#> 2        42.96346      30.16725     56.605602
#> 3        34.87141      28.60429     41.461819
#> 4        41.65088      34.82491     48.714911
#> 5        37.88513      28.81459     47.327183
#> 6        39.98055      32.96272     46.901771
#> 7        39.54465      30.73839     49.034735
#> 8        37.84685      29.89618     46.136840
#> 9        23.16909      15.52742     32.167790
#> 10       47.41763      40.79040     54.003459
#> 11       46.25578      37.29341     55.486637
#> 12       39.60942      32.40247     47.010114
#> 13       42.16994      36.05525     48.014521
#> 14       32.69376      25.32539     40.743616
#> 15       39.59220      31.92089     47.639160
#> 16       28.69969      20.56018     37.765601
#> 17       46.21862      38.06191     54.820900
#> 18       29.79790      21.84763     38.184801
#> 19       40.80147      34.01954     47.731448
#> 20       38.81229      32.47121     45.134050
#> 21       44.19449      38.34391     49.863610
#> 22       43.03578      41.39910     43.516761
#> 23       39.91552      34.03111     45.828909
#> 24       32.73711      26.11832     39.822330
#> 25       31.19976      24.35551     39.070132
#> 26       38.70954      31.51130     46.162112
#> 27       42.26111      34.55727     49.962592
#> 28       35.93959      28.74806     43.443961
#> 29       41.10020      34.17442     48.287123
#> 30       45.23072      38.71904     51.753689
#> 31       40.00692      34.08098     46.118968
#> 32       36.07692      29.57571     42.875228
#> 33       36.68642      30.16876     43.661649
#> 34       36.80163      30.28742     43.594493
#> 35       45.54335      37.30655     54.136209
#> 36       43.46914      38.25682     48.235586
#> 37       31.82500      22.66728     42.033232
#> 38       28.22226      18.90537     39.444559
#> 39       35.34858      28.10651     43.220210
#> 40        0.00000       0.00000      8.496889
#> 41       25.17574      17.29056     34.215661
#> 42       39.46089      32.50697     46.741205
#> 43       43.76537      37.09142     50.411768
#> 44       30.50019      21.48018     40.485977
#> 45       46.30977      39.02693     53.452473
#> 46       32.25925      24.99587     39.958538
#> 47       42.70247      35.25502     50.399605
#> 48       28.35869      20.62415     37.062916
#> 49        0.00000       0.00000     20.083134
#> 50       20.93428      13.75549     29.536098
#> 51       30.11671      21.39002     39.776565
#> 52       34.09906      25.77339     43.041449
#> 53       40.72475      32.40145     49.234912
#> 54       39.44729      32.89185     46.283816
#> 55       33.11384      26.33305     40.377536
#> 56       42.32394      34.72332     50.171511
#> 57       42.23843      34.61668     50.573791
#> 58       34.08837      27.09223     41.531194
#> 59       39.92598      33.00018     47.103519
#> 60       36.01936      29.03467     43.375715
#> 61       47.35991      42.70740     51.433610
#> 62       44.80534      38.32684     50.997671
#> 63       42.97603      34.22036     52.255017
#> 64       42.63042      33.95337     51.711821
#> 65       30.90611      22.09495     40.368445
#> 66       38.72889      31.89834     45.492734
#> 67       32.50245      22.34867     44.402502
#> 68       39.26238      33.24759     45.398668
#> 69       34.85361      27.58244     42.573134
#> 70       33.28034      25.39379     41.973045
#> 71       36.05373      27.44841     45.300900
#> 72       19.30831      10.76798     30.302661
#> 73       32.26173      25.19733     39.790719
#> 74       26.52065      18.34504     36.027691
#> 75       48.60084      41.82453     55.076444
#> 76       42.36079      34.49847     50.417850
#> 77       30.19135      21.49342     40.209699
#> 78       41.93362      35.87922     47.954484
#> 79       38.17236      30.64560     46.149146
#> 80       25.78715      17.89112     34.768312
#> 81        0.00000       0.00000     44.287034
#> 82       42.09702      34.17498     50.430559
#> 83       42.51508      37.02073     47.686775
#> 84       43.63949      37.13631     50.100950
#> 85       44.74284      37.76760     51.769614
#> 86       28.76993      20.63797     37.848323
#> 87       36.12595      28.62036     44.009834
#> 88       42.09016      35.68355     48.342476
#> 89       34.60636      27.27913     42.408191
#> 90       43.59159      35.31038     52.172657
#> 91       41.69611      35.66251     47.670684
#> 92       39.12839      32.41103     45.776640
#> 93       30.75639      22.23575     40.926068
#> 94       27.66744      20.14998     35.705094
#> 95       43.46892      37.56324     49.251572
#> 96       44.26188      42.56234     44.745191
#> 97       42.08025      35.85165     48.311859
#> 98       35.50473      27.03179     45.044471
#> 99       36.94984      29.92630     44.372109
#> 100      42.73608      35.43082     50.157780
#>