Get confidence intervals for provider effects or standardized measures from a fitted logis_re object

Provide confidence intervals for provider effects or standardized measures from a random effect logistic model.

Usage

# S3 method for class 'logis_re'
confint(
  object,
  parm,
  level = 0.95,
  option = "SM",
  measure = c("rate", "ratio"),
  stdz = "indirect",
  alternative = "two.sided",
  ...
)

Arguments

object

a model fitted from logis_re.

parm

specify a subset of providers for which confidence intervals are given. By default, all providers are included. The class of parm should match the class of the provider IDs.

level

the confidence level. The default value is 0.95.

option

a character string specifying whether the confidence intervals should be provided for provider effects or standardized measures:

"alpha" provider effect.

"SM" standardized measures.

measure

a character string or a vector indicating whether the output measure is "ratio" or "rate" if option = "SM". Both "rate" and "ratio" will be provided by default.

stdz

a character string or a vector specifying the standardization method if option includes "SM". See stdz argument in SM_output.linear_re.

alternative

a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater", or "less". Note that "alpha" for argument option only supports "two.sided".

...

additional arguments that can be passed to the function.

Value

A list of data frames containing the confidence intervals based on the values of option and stdz.

CI.alpha

Confidence intervals for provider effects if option includes "alpha".

CI.indirect

Confidence intervals for indirect standardized differences if option includes "SM" and stdz includes "indirect".

CI.direct

Confidence intervals for direct standardized differences if option includes "SM" and stdz includes "direct".

Examples

data(ExampleDataBinary)
outcome <- ExampleDataBinary$Y
ProvID <- ExampleDataBinary$ProvID
covar <- ExampleDataBinary$Z
fit_re <- logis_re(Y = outcome, Z = covar, ProvID = ProvID)
#> Input format: Y, Z, and ProvID.
confint(fit_re)
#> $CI.indirect_ratio
#>     Indirect.Ratio Ratio.Lower Ratio.Upper
#> 1        0.9461038   0.8166696   1.0775785
#> 2        1.0787176   0.8266316   1.3509527
#> 3        0.9400063   0.8020679   1.0854291
#> 4        1.0653672   0.9168257   1.2205995
#> 5        0.9987896   0.8080054   1.1978990
#> 6        1.0326903   0.8831402   1.1807069
#> 7        1.0295765   0.8391357   1.2369977
#> 8        0.9969787   0.8266234   1.1752197
#> 9        0.7310625   0.5547865   0.9335291
#> 10       1.1559239   1.0186377   1.2957306
#> 11       1.1513940   0.9586925   1.3534230
#> 12       1.0272776   0.8727429   1.1869114
#> 13       1.0659543   0.9385403   1.1881813
#> 14       0.9072552   0.7493832   1.0789126
#> 15       1.0284386   0.8634648   1.2026720
#> 16       0.8368009   0.6587867   1.0335479
#> 17       1.1505361   0.9753888   1.3397750
#> 18       0.8616247   0.6929460   1.0375824
#> 19       1.0537609   0.9018239   1.2100342
#> 20       1.0129447   0.8736556   1.1518772
#> 21       1.1043917   0.9803733   1.2261455
#> 22       1.0574039   1.0192613   1.0847512
#> 23       1.0278070   0.9057104   1.1507400
#> 24       0.9018841   0.7562567   1.0581106
#> 25       0.8746183   0.7222845   1.0506907
#> 26       1.0095920   0.8593705   1.1654891
#> 27       1.0708732   0.9101558   1.2329351
#> 28       0.9612353   0.8039374   1.1256362
#> 29       1.0516988   0.9050124   1.2046927
#> 30       1.1310946   0.9893368   1.2745319
#> 31       1.0281327   0.9073688   1.1533687
#> 32       0.9626679   0.8213761   1.1107916
#> 33       0.9752342   0.8399320   1.1204406
#> 34       0.9765691   0.8344976   1.1251239
#> 35       1.1257201   0.9552898   1.3045385
#> 36       1.0848737   0.9762584   1.1844440
#> 37       0.8966251   0.7012939   1.1131570
#> 38       0.8333818   0.6267944   1.0782120
#> 39       0.9519421   0.7956057   1.1223155
#> 40       0.6917523   0.4838175   0.9823874
#> 41       0.7898787   0.6177359   0.9789330
#> 42       1.0243667   0.8753752   1.1812538
#> 43       1.0958247   0.9555165   1.2359668
#> 44       0.8720864   0.6769959   1.0866518
#> 45       1.1493913   0.9917336   1.3057765
#> 46       0.8987681   0.7429220   1.0634761
#> 47       1.0814394   0.9234054   1.2463802
#> 48       0.8283659   0.6572666   1.0195584
#> 49       0.7932016   0.5312365   1.1546570
#> 50       0.7129017   0.5510319   0.9018570
#> 51       0.8669821   0.6792316   1.0729878
#> 52       0.9338524   0.7577814   1.1224609
#> 53       1.0470292   0.8715055   1.2278014
#> 54       1.0195648   0.8870492   1.1584716
#> 55       0.9102634   0.7616757   1.0691060
#> 56       1.0819407   0.9142257   1.2568362
#> 57       1.0725598   0.9124628   1.2502529
#> 58       0.9281800   0.7756123   1.0904425
#> 59       1.0318310   0.8845132   1.1855691
#> 60       0.9631140   0.8137316   1.1209518
#> 61       1.1651478   1.0615454   1.2577753
#> 62       1.1154883   0.9776139   1.2480308
#> 63       1.0803970   0.9022076   1.2706975
#> 64       1.0798398   0.8977886   1.2723232
#> 65       0.8878000   0.7065769   1.0795961
#> 66       1.0115330   0.8627720   1.1587841
#> 67       0.9177792   0.7124482   1.1565826
#> 68       1.0184643   0.8903398   1.1496207
#> 69       0.9420426   0.7830151   1.1109811
#> 70       0.9198589   0.7533020   1.1031269
#> 71       0.9667711   0.7846122   1.1628624
#> 72       0.6984532   0.5018949   0.9379784
#> 73       0.8933515   0.7369523   1.0595700
#> 74       0.8047422   0.6258156   1.0117340
#> 75       1.2037113   1.0516442   1.3512664
#> 76       1.0762684   0.9091183   1.2491031
#> 77       0.8706838   0.6851209   1.0828299
#> 78       1.0686850   0.9374246   1.2002539
#> 79       1.0016882   0.8422959   1.1714695
#> 80       0.7949570   0.6221143   0.9861896
#> 81       0.7750384   0.3697166   1.5462937
#> 82       1.0749287   0.9036995   1.2571988
#> 83       1.0798825   0.9588486   1.1939284
#> 84       1.0994113   0.9591325   1.2400378
#> 85       1.1199411   0.9697523   1.2726253
#> 86       0.8496973   0.6792562   1.0389106
#> 87       0.9656771   0.8026670   1.1371800
#> 88       1.0773306   0.9338145   1.2179646
#> 89       0.9390843   0.7820825   1.1066239
#> 90       1.0977167   0.9225807   1.2815632
#> 91       1.0615548   0.9326217   1.1894167
#> 92       1.0172649   0.8738242   1.1591957
#> 93       0.8767679   0.6920784   1.0969399
#> 94       0.8407536   0.6869406   0.9996060
#> 95       1.1007761   0.9696311   1.2293436
#> 96       1.0778057   1.0366482   1.1089599
#> 97       1.0743646   0.9371026   1.2123837
#> 98       0.9601327   0.7872421   1.1543034
#> 99       0.9796348   0.8258286   1.1430846
#> 100      1.0841601   0.9271759   1.2454284
#> 
#> $CI.indirect_rate
#>     Indirect.Rate Rate.Lower Rate.Upper
#> 1        36.68177   31.66342   41.77923
#> 2        41.82339   32.04967   52.37833
#> 3        36.44536   31.09730   42.08361
#> 4        41.30578   35.54662   47.32435
#> 5        38.72447   31.32750   46.44422
#> 6        40.03885   34.24058   45.77766
#> 7        39.91812   32.53446   47.96013
#> 8        38.65426   32.04935   45.56491
#> 9        28.34432   21.50985   36.19423
#> 10       44.81679   39.49401   50.23729
#> 11       44.64116   37.16985   52.47411
#> 12       39.82899   33.83746   46.01822
#> 13       41.32854   36.38852   46.06745
#> 14       35.17556   29.05463   41.83095
#> 15       39.87400   33.47774   46.62928
#> 16       32.44394   25.54208   40.07210
#> 17       44.60790   37.81719   51.94495
#> 18       33.40639   26.86649   40.22852
#> 19       40.85579   34.96497   46.91472
#> 20       39.27329   33.87285   44.65989
#> 21       42.81881   38.01044   47.53938
#> 22       40.99703   39.51819   42.05732
#> 23       39.84952   35.11566   44.61580
#> 24       34.96731   29.32113   41.02443
#> 25       33.91018   28.00398   40.73675
#> 26       39.14329   33.31900   45.18764
#> 27       41.51925   35.28801   47.80262
#> 28       37.26844   31.16978   43.64249
#> 29       40.77584   35.08860   46.70762
#> 30       43.85412   38.35797   49.41539
#> 31       39.86215   35.17996   44.71772
#> 32       37.32398   31.84590   43.06695
#> 33       37.81120   32.56534   43.44105
#> 34       37.86295   32.35464   43.62263
#> 35       43.64574   37.03792   50.57878
#> 36       42.06207   37.85090   45.92255
#> 37       34.76341   27.19015   43.15866
#> 38       32.31138   24.30170   41.80379
#> 39       36.90813   30.84675   43.51374
#> 40       26.82020   18.75828   38.08853
#> 41       30.62470   23.95049   37.95460
#> 42       39.71613   33.93952   45.79886
#> 43       42.48666   37.04671   47.92016
#> 44       33.81201   26.24808   42.13101
#> 45       44.56351   38.45090   50.62678
#> 46       34.84650   28.80413   41.23246
#> 47       41.92892   35.80172   48.32391
#> 48       32.11690   25.48315   39.52970
#> 49       30.75354   20.59678   44.76767
#> 50       27.64020   21.36428   34.96626
#> 51       33.61411   26.33476   41.60124
#> 52       36.20676   29.38025   43.51938
#> 53       40.59479   33.78949   47.60358
#> 54       39.52996   34.39214   44.91556
#> 55       35.29219   29.53123   41.45073
#> 56       41.94835   35.44581   48.72930
#> 57       41.58465   35.37746   48.47406
#> 58       35.98684   30.07157   42.27798
#> 59       40.00553   34.29381   45.96617
#> 60       37.34128   31.54951   43.46087
#> 61       45.17441   41.15760   48.76571
#> 62       43.24904   37.90346   48.38790
#> 63       41.88850   34.97985   49.26672
#> 64       41.86690   34.80852   49.32975
#> 65       34.42125   27.39497   41.85745
#> 66       39.21855   33.45088   44.92768
#> 67       35.58359   27.62261   44.84233
#> 68       39.48729   34.51972   44.57240
#> 69       36.52431   30.35859   43.07429
#> 70       35.66422   29.20657   42.76977
#> 71       37.48307   30.42051   45.08580
#> 72       27.08001   19.45917   36.36673
#> 73       34.63649   28.57267   41.08101
#> 74       31.20098   24.26375   39.22634
#> 75       46.66957   40.77372   52.39049
#> 76       41.72843   35.24779   48.42948
#> 77       33.75763   26.56310   41.98283
#> 78       41.43441   36.34527   46.53553
#> 79       38.83685   32.65699   45.41951
#> 80       30.82160   24.12024   38.23595
#> 81       30.04932   14.33443   59.95197
#> 82       41.67649   35.03770   48.74336
#> 83       41.86856   37.17590   46.29027
#> 84       42.62571   37.18691   48.07800
#> 85       43.42169   37.59865   49.34146
#> 86       32.94395   26.33571   40.28002
#> 87       37.44065   31.12052   44.09006
#> 88       41.76961   36.20529   47.22219
#> 89       36.40961   30.32243   42.90536
#> 90       42.56001   35.76974   49.68800
#> 91       41.15796   36.15905   46.11535
#> 92       39.44079   33.87939   44.94364
#> 93       33.99352   26.83285   42.52990
#> 94       32.59719   26.63365   38.75612
#> 95       42.67863   37.59395   47.66337
#> 96       41.78804   40.19230   42.99593
#> 97       41.65462   36.33278   47.00581
#> 98       37.22569   30.52248   44.75396
#> 99       37.98181   32.01853   44.31899
#> 100      42.03440   35.94791   48.28700
#>