Data quality check function

Conduct data quality check including checking missingness, variation, correlation and VIF of variables.

Usage

data_check(Y, Z, ProvID)

Arguments

Y

a numeric vector indicating the outcome variable.

Z

a matrix or data frame representing covariates.

ProvID

a numeric vector representing the provider identifier.

Value

No return value, called for side effects.

Details

The function performs the following checks:

Missingness: Checks for any missing values in the dataset and provides a summary of missing data.

Variation: Identifies covariates with zero or near-zero variance which might affect model stability.

Correlation: Analyzes pairwise correlation among covariates and highlights highly correlated pairs.

VIF: Computes the Variable Inflation Factors to identify covariates with potential multicollinearity issues.

If issues arise when using the model functions logis_fe, linear_fe and linear_re, this function can be called for data quality checking purposes.

Examples

data(ExampleDataBinary)
outcome = ExampleDataBinary$Y
covar = ExampleDataBinary$Z
ProvID = ExampleDataBinary$ProvID
data_check(outcome, covar, ProvID)
#> Checking missingness of variables ... 
#> Missing values NOT found. Checking missingness of variables completed!
#> Checking variation in covariates ... 
#> Checking variation in covariates completed!
#> Checking pairwise correlation among covariates ... 
#> Checking pairwise correlation among covariates completed!
#> Checking VIF of covariates ... 
#> Checking VIF of covariates completed!