Get Funnel Plot for institutional comparisons

Creates a funnel plot from a logistic fixed-effect model to compare provider performance.

Usage

# S3 method for logis_fe
plot(
  fit,
  null = "median",
  test = "score",
  target = 1,
  alpha = c(0.05, 0.01),
  labels = c("lower", "expected", "higher"),
  point_colors = c("#E69F00", "#56B4E9", "#009E73"),
  point_shapes = c(15, 17, 19),
  point_size = 2,
  point_alpha = 0.8,
  line_size = 0.8,
  target_line_type = "longdash"
)

Arguments

fit

a model fitted from logis_fe.

null

a character string or a number specifying null hypotheses of fixed provider effects. The default is "median".

test

a character string specifying the type of testing method to be conducted. The default is "score".

target

a numeric value representing the target outcome. The default value is 1.

labels

a vector of labels for the plot.

point_colors

a vector of colors representing different provider flags. The default is c("#E69F00", "#56B4E9", "#009E73").

point_shapes

a vector of shapes representing different provider flags. The default is c(15, 17, 19).

point_size

size of the points. The default is 2.

point_alpha

alpha of the points. The default is 0.8.

line_size

size of all lines, including control limits and the target line. The default is 0.8.

target_line_type

line type for the target line. The default is "longdash".

alphas

a vector of significance levels. The default is alpha = c(0.05, 0.01).

Value

A ggplot object representing the funnel plot.

Details

This function generates a funnel plot from a logistic fixed-effect model. Currently, it only supports the indirect standardized ratio. The parameter alpha is a vector used to calculate control limits at different significance levels. The first value in the vector is used as the significance level for flagging each provider, utilizing the test.logis_fe function.

See also

Examples

data(data_FE)
#> Warning: data set 'data_FE' not found
fit_fe <- logis_fe(Y = data_FE$Y, Z = data_FE$Z, ID = data_FE$ID, message = FALSE)
#> Error: object 'data_FE' not found
plot(fit_fe)
#> Error: object 'fit_fe' not found