Result Summaries of Covariate Estimates from a fitted linear_fe or linear_re object

Provide the summary statistics for the covariate estimates for a fixed/random effect linear model.

Usage

# S3 method for class 'linear_fe'
summary(object, parm, level = 0.95, null = 0, ...)

# S3 method for class 'linear_re'
summary(object, parm, level = 0.95, null = 0, ...)

Arguments

object

a model fitted from linear_fe or linear_re.

parm

Specifies a subset of covariates for which the result summaries should be output. By default, all covariates are included.

level

the confidence level during the hypothesis test, meaning a significance level of \(1 - \text{level}\). The default value is 0.95.

null

a number defining the null hypothesis for the covariate estimates. The default value is 0.

...

additional arguments that can be passed to the function.

Value

A data frame containing summary statistics for covariate estimates, with the following columns:

Estimate

the estimates of covariate coefficients.

Std.Error

the standard error of the estimate.

Stat

the test statistic.

p value

the p-value for the hypothesis test.

CI.upper

the lower bound of the confidence interval.

CI.lower

the upper bound of the confidence interval.

Examples

data(ExampleDataLinear)
outcome <- ExampleDataLinear$Y
covar <- ExampleDataLinear$Z
ProvID <- ExampleDataLinear$ProvID
fit_fe <- linear_fe(Y = outcome, Z = covar, ProvID = ProvID)
#> Input format: Y, Z, and ProvID.
summary(fit_fe)
#>     Estimate  Std.Error     Stat p value  CI.Lower CI.Upper
#> z1 1.0150364 0.01146089 88.56523 < 1e-10 0.9925700 1.037503
#> z2 0.9908107 0.01146506 86.42003 < 1e-10 0.9683361 1.013285
#> z3 0.9982308 0.01137525 87.75459 < 1e-10 0.9759322 1.020529
#> z4 0.9900928 0.01165625 84.94096 < 1e-10 0.9672435 1.012942
#> z5 0.9961593 0.01156610 86.12752 < 1e-10 0.9734867 1.018832

data(ExampleDataLinear)
outcome <- ExampleDataLinear$Y
covar <- ExampleDataLinear$Z
ProvID <- ExampleDataLinear$ProvID
fit_re <- linear_fe(Y = outcome, Z = covar, ProvID = ProvID)
#> Input format: Y, Z, and ProvID.
summary(fit_re)
#>     Estimate  Std.Error     Stat p value  CI.Lower CI.Upper
#> z1 1.0150364 0.01146089 88.56523 < 1e-10 0.9925700 1.037503
#> z2 0.9908107 0.01146506 86.42003 < 1e-10 0.9683361 1.013285
#> z3 0.9982308 0.01137525 87.75459 < 1e-10 0.9759322 1.020529
#> z4 0.9900928 0.01165625 84.94096 < 1e-10 0.9672435 1.012942
#> z5 0.9961593 0.01156610 86.12752 < 1e-10 0.9734867 1.018832