Result Summaries of Covariate Estimates from a fitted logis_re object

Provide the summary statistics for the covariate estimates for a random effect logistic model.

Usage

# S3 method for class 'logis_re'
summary(object, parm, level = 0.95, null = 0, ...)

Arguments

object

a model fitted from logis_re.

parm

specifies a subset of covariates for which the result summaries should be output. By default, all covariates are included.

level

the confidence level during the hypothesis test, meaning a significance level of \(1 - \text{level}\). The default value is 0.95.

null

a number defining the null hypothesis for the covariate estimates. The default value is 0.

...

additional arguments that can be passed to the function.

Value

A data frame containing summary statistics for covariate estimates, with the following columns:

Estimate

the estimates of covariate coefficients.

Std.Error

the standard error of the estimate.

Stat

the test statistic.

p value

the p-value for the hypothesis test.

CI.upper

the lower bound of the confidence interval.

CI.lower

the upper bound of the confidence interval.

Examples

data(ExampleDataBinary)
outcome <- ExampleDataBinary$Y
covar <- ExampleDataBinary$Z
ProvID <- ExampleDataBinary$ProvID
fit_re <- logis_re(Y = outcome, Z = covar, ProvID = ProvID)
#> Input format: Y, Z, and ProvID.
summary(fit_re)
#>               Estimate  Std.Error      Stat p value   CI.Lower  CI.Upper
#> (Intercept) -0.9847415 0.06091519 -16.16578       2 -1.1041331 -0.865350
#> z1           1.0873566 0.04068289  26.72762  <1e-10  1.0076196  1.167094
#> z2           1.0182187 0.04037506  25.21900  <1e-10  0.9390850  1.097352
#> z3           0.9948328 0.03949206  25.19071  <1e-10  0.9174298  1.072236
#> z4           1.0346604 0.04034811  25.64334  <1e-10  0.9555795  1.113741
#> z5           1.0463216 0.04092690  25.56562  <1e-10  0.9661063  1.126537