Conduct hypothesis testing for provider effects from a fitted logis_re object

Conduct hypothesis tests on provider effects and identify outlying providers for a random effect logistic model.

Usage

# S3 method for class 'logis_re'
test(fit, parm, level = 0.95, null = 0, alternative = "two.sided", ...)

Arguments

fit

a model fitted from logis_re.

parm

specifies a subset of providers for which confidence intervals are to be given. By default, all providers are included. The class of parm should match the class of the provider IDs.

level

the confidence level during the hypothesis test, meaning a significance level of \(1 - \text{level}\). The default value is 0.95.

null

a number defining the null hypothesis for the provider effects. The default value is 0.

alternative

a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater", or "less".

...

additional arguments that can be passed to the function.

Value

A data frame containing the results of the hypothesis test, with the following columns:

flag

a flagging indicator where 1 means statistically higher than expected and -1 means statistically lower than expected.

p-value

the p-value of the hypothesis test.

stat

the test statistic.

Std.Error

the standard error of the provider effect estimate.

Details

The function identifies outlying providers based on hypothesis test results. For two-sided tests, 1 indicates performance significantly higher than expected, -1 indicates lower, For one-sided tests, 1 (right-tailed) or -1 (left-tailed) flags are used. Providers whose performance falls within the central range are flagged as 0. Outlying providers are determined by the test statistic falling beyond the threshold based on the significance level \(1 - \text{level}\).

Examples

data(ExampleDataBinary)
outcome <- ExampleDataBinary$Y
ProvID <- ExampleDataBinary$ProvID
covar <- ExampleDataBinary$Z
fit_re <- logis_re(Y = outcome, Z = covar, ProvID = ProvID)
#> Input format: Y, Z, and ProvID.
test(fit_re)
#>     flag      p value        stat Std.Error
#> 1      0 0.4196420524 -0.80704241 0.2828733
#> 2      0 0.5547024534  0.59072834 0.3048979
#> 3      0 0.4127679229 -0.81903241 0.2547667
#> 4      0 0.3962562702  0.84832631 0.2543625
#> 5      0 0.9903824508 -0.01205410 0.2794248
#> 6      0 0.6692179953  0.42722151 0.2701339
#> 7      0 0.7706118277  0.29157469 0.2613230
#> 8      0 0.9730289610 -0.03380962 0.2682870
#> 9     -1 0.0104204961 -2.56155426 0.2670380
#> 10     1 0.0255899959  2.23237792 0.2773311
#> 11     0 0.1265620763  1.52780051 0.2605441
#> 12     0 0.7346437494  0.33895484 0.2656730
#> 13     0 0.3069866112  1.02156546 0.2825115
#> 14     0 0.2823859398 -1.07497496 0.2767873
#> 15     0 0.7421079108  0.32906321 0.2644456
#> 16     0 0.1016130042 -1.63708359 0.2739273
#> 17     0 0.0946198905  1.67151555 0.2547047
#> 18     0 0.1220635999 -1.54616966 0.2830245
#> 19     0 0.4937153533  0.68441150 0.2394212
#> 20     0 0.8558668333  0.18163799 0.2533804
#> 21     0 0.0991212515  1.64912880 0.2714367
#> 22     1 0.0055602853  2.77264435 0.3506551
#> 23     0 0.6568239154  0.44430247 0.2812819
#> 24     0 0.2131520708 -1.24494472 0.2588448
#> 25     0 0.1551850780 -1.42145310 0.2608820
#> 26     0 0.9021919328  0.12289287 0.2796284
#> 27     0 0.3909715534  0.85785678 0.2793539
#> 28     0 0.6397549794 -0.46804141 0.2581519
#> 29     0 0.4961824916  0.68050858 0.2696851
#> 30     0 0.0701968823  1.81063815 0.2547598
#> 31     0 0.6537381040  0.44857523 0.2878787
#> 32     0 0.6160709195 -0.50142661 0.2604814
#> 33     0 0.7313482556 -0.34333260 0.2838420
#> 34     0 0.7539017622 -0.31349876 0.2585329
#> 35     0 0.1510559015  1.43581180 0.2795602
#> 36     0 0.1224181875  1.54470272 0.2888328
#> 37     0 0.3389291816 -0.95628319 0.2823548
#> 38     0 0.1730149206 -1.36257998 0.2789475
#> 39     0 0.5706738004 -0.56705944 0.2677016
#> 40    -1 0.0393181849 -2.06084128 0.4009507
#> 41    -1 0.0297799772 -2.17300452 0.2918425
#> 42     0 0.7543855676  0.31286192 0.2653412
#> 43     0 0.1810797587  1.33743784 0.2758533
#> 44     0 0.2353518871 -1.18668502 0.2775724
#> 45     0 0.0633716982  1.85657462 0.2610245
#> 46     0 0.2244812054 -1.21469824 0.2734383
#> 47     0 0.3191172316  0.99627359 0.2684920
#> 48     0 0.0771435268 -1.76750599 0.2700515
#> 49     0 0.2285028361 -1.20422446 0.4228710
#> 50    -1 0.0039878065 -2.87912467 0.2917094
#> 51     0 0.1996264933 -1.28261642 0.2811697
#> 52     0 0.4841838116 -0.69958932 0.2807298
#> 53     0 0.6060658318  0.51569731 0.2748806
#> 54     0 0.7780089075  0.28191471 0.2926346
#> 55     0 0.2630604696 -1.11918677 0.2622536
#> 56     0 0.3454180129  0.94351409 0.2475795
#> 57     0 0.3894558907  0.86060453 0.2694271
#> 58     0 0.3792140584 -0.87934534 0.2630616
#> 59     0 0.6785958605  0.41437996 0.2686551
#> 60     0 0.6404614915 -0.46705365 0.2694790
#> 61     1 0.0022779378  3.05137776 0.2658253
#> 62     0 0.0999569311  1.64506246 0.2735004
#> 63     0 0.3860734878  0.86676006 0.2880330
#> 64     0 0.3993842376  0.84272147 0.2747748
#> 65     0 0.2477259806 -1.15589042 0.2983926
#> 66     0 0.8791852030  0.15200220 0.2591200
#> 67     0 0.4793809780 -0.70729853 0.3082824
#> 68     0 0.7803781130  0.27882632 0.2679748
#> 69     0 0.4951942507 -0.68207069 0.2603592
#> 70     0 0.3803657497 -0.87722259 0.2831437
#> 71     0 0.7331626190 -0.34092158 0.2774745
#> 72    -1 0.0156627299 -2.41668116 0.3011520
#> 73     0 0.2050457556 -1.26730639 0.2569424
#> 74     0 0.0634968150 -1.85569661 0.2833437
#> 75     1 0.0085868131  2.62808074 0.2430865
#> 76     0 0.3772965357  0.88288847 0.2687735
#> 77     0 0.2217876276 -1.22178848 0.2878268
#> 78     0 0.3077067542  1.02004577 0.2582064
#> 79     0 0.9839657543  0.02009730 0.2727857
#> 80    -1 0.0360577005 -2.09627616 0.2861312
#> 81     0 0.4802658583 -0.70587503 0.4189495
#> 82     0 0.4015626484  0.83883368 0.2584268
#> 83     0 0.1925795625  1.30298594 0.2553479
#> 84     0 0.1661090905  1.38481485 0.2612584
#> 85     0 0.1187115783  1.56020443 0.2613430
#> 86     0 0.1157734817 -1.57276397 0.2945649
#> 87     0 0.6905946246 -0.39804824 0.2626102
#> 88     0 0.2908105623  1.05634513 0.2420262
#> 89     0 0.4676862241 -0.72624886 0.2710364
#> 90     0 0.2814179229  1.07713956 0.2694723
#> 91     0 0.3492272658  0.93608919 0.2674803
#> 92     0 0.8133638592  0.23608868 0.2693033
#> 93     0 0.2561942663 -1.13543222 0.2808371
#> 94    -1 0.0494372204 -1.96480147 0.3059605
#> 95     0 0.1314942961  1.50823538 0.2478016
#> 96     1 0.0006298054  3.41844285 0.3399443
#> 97     0 0.2902336522  1.05760919 0.2496929
#> 98     0 0.6754261450 -0.41871266 0.2974836
#> 99     0 0.8024027857 -0.25023867 0.2556061
#> 100    0 0.2984244090  1.03981812 0.2626159