Example data with binary outcomes

A simulated data set containing 7994 observations, 5 continuous covariates and 100 providers.

Usage

data(ExampleDataBinary)

Format

A list containing the following elements:

Y

a vector representing binary outcomes with 0 or 1. Generated from a Bernoulli distribution with the probability of success (\(\boldsymbol{\mu}\)), which is determined by applying the logistic function to the linear combination of provider effects and covariates.

ProvID

a vector representing identifier for each provider. The number of individuals per provider is generated from a Poisson distribution with mean 80, with a minimum value of 11.

Z

a data frame containing 5 continuous variables. Generated from a multivariate normal distribution, where the mean is calculated as \((\gamma_i - \mu_{\gamma}) \cdot \rho / \sigma_{\gamma}\) with \(\mu_{\gamma} = \log(4/11)\) and \(\sigma_{\gamma} = 0.4\), and the correlation between covariates and provider effects is 0.1.

Examples

data(ExampleDataBinary)
head(ExampleDataBinary$Y)
#> [1] 1 1 0 0 1 0
head(ExampleDataBinary$ProvID)
#> [1] 1 1 1 1 1 1
head(ExampleDataBinary$Z)
#>           z1         z2         z3          z4          z5
#> 1  1.1588563  1.5038359 0.51678022  0.64381150  0.07779473
#> 2 -0.4065060  0.0988566 0.42928104 -0.02097142  1.87616869
#> 3 -0.9162474 -1.0128945 0.68255389 -0.21407813  1.14789766
#> 4  0.1946768  1.4542295 1.19780422 -0.34841782 -0.28220979
#> 5  1.3290611  0.4604717 0.09888742 -0.83724543  0.66339830
#> 6 -0.1260386 -2.2552490 0.48275904 -0.72250452  0.61493456