Example data with continuous outcomes

A simulated data set containing 7901 observations, 5 continuous covariates and 100 providers.

Usage

data(ExampleDataLinear)

Format

A list containing the following elements:

Y

a vector representing continuous outcomes. Generated as a linear combination of provider-specific effects, covariates, and a random error term (\(\epsilon\)), where \(\epsilon\) follows a normal distribution with mean 0 and standard deviation 1.

ProvID

a vector representing identifier for each provider. The number of individuals per provider is generated from a Poisson distribution with mean 80.

Z

a data frame containing 5 continuous variables. Generated from a multivariate normal distribution, where the mean is calculated as \((\gamma_i - \mu_{\gamma}) \cdot \rho / \sigma_{\gamma}\) with \(\mu_{\gamma} = \log(4/11)\) and \(\sigma_{\gamma} = 0.4\), and the correlation between covariates and provider effects is 0.1.

Examples

data(ExampleDataLinear)
head(ExampleDataLinear$Y)
#> [1]  2.341395  1.726385 -6.085487 -4.459568 -5.392451 -3.747529
head(ExampleDataLinear$ProvID)
#> [1] 1 1 1 1 1 1
head(ExampleDataLinear$Z)
#>           z1         z2         z3         z4          z5
#> 1  1.9076523  0.3966139  0.2295083  0.9965029  1.63538482
#> 2  1.2768349 -0.2514242  0.9709646  1.6304805 -0.51758434
#> 3 -0.3510678 -0.5429142 -0.3862969 -1.3255520 -1.69976484
#> 4 -0.6409343 -0.8019597 -1.2999300 -1.6362957 -0.42451428
#> 5 -1.2625848 -0.4706832 -0.8130136  0.1509436 -0.95600474
#> 6 -0.2724952 -0.6752503 -2.1985430 -0.6236876 -0.05059055