Calculate direct/indirect standardized ratios/rates from a fitted logis_fe object

Provide direct/indirect standardized ratios/rates for a fixed effect logistic model.

Usage

# S3 method for logis_fe
SM_output(
  fit,
  parm,
  stdz = "indirect",
  measure = c("rate", "ratio"),
  null = "median",
  threads = 2
)

Arguments

fit

a model fitted from logis_fe.

parm

specifies a subset of providers for which confidence intervals are to be given. By default, all providers are included. The class of parm should match the class of the provider IDs.

stdz

a character string or a vector specifying the standardization method(s). The possible values are:

"indirect" (default) indirect standardization method.

"direct" direct standardization method.

c("indirect", "direct") outputs both direct and indirect standardized measures.

measure

a character string or a vector indicating whether the output measure is "ratio" or "rate"

"rate" output the standardized rate. The "rate" has been restricted to 0% - 100%.

"ratio" output the standardized ratio.

c("ratio", "rate") (default) output both the ratio and rate.

null

if "stdz = indirect", a character string or a number defining the population norm. The default is "median".

threads

an integer specifying the number of threads to use. The default value is 2.

Value

A list contains standardized measures, as well as the observed and expected outcomes used for calculation, depending on the user's choice of standardization method (stdz) and measure type (measure).

indirect.ratio

standardization ratio using indirect method if stdz includes "indirect" and measure includes "ratio".

direct.ratio

standardization ratio using direct method if stdz includes "direct" and measure includes "ratio".

indirect.rate

standardization rate using indirect method if stdz includes "indirect" and measure includes "rate".

direct.rate

standardization rate using direct method if stdz includes "direct" and measure includes "rate".

OE

a list of data frames containing the observed and expected outcomes used for calculating standardized measures.

References

He, K. (2019). Indirect and direct standardization for evaluating transplant centers. Journal of Hospital Administration, 8(1), 9-14.

Examples

data(ExampleDataBinary)
outcome = ExampleDataBinary$Y
covar = ExampleDataBinary$Z
ID = ExampleDataBinary$ID
fit_fe <- logis_fe(Y = outcome, Z = covar, ID = ID, message = FALSE)
SR <- SM_output(fit_fe, stdz = "direct", measure = "rate")
SR$direct.rate
#>     Direct_standardized.rate
#> 1                 34.4982348
#> 2                 42.0717647
#> 3                 35.1841930
#> 4                 42.0355239
#> 5                 38.4468700
#> 6                 40.4679520
#> 7                 39.7796647
#> 8                 38.3561364
#> 9                 27.5688752
#> 10                49.5923738
#> 11                45.1347396
#> 12                40.0121255
#> 13                43.7234206
#> 14                33.4452622
#> 15                39.9364904
#> 16                31.0598534
#> 17                45.4819829
#> 18                30.9478859
#> 19                41.0429442
#> 20                39.2278237
#> 21                46.2815438
#> 22                73.2207221
#> 23                40.7257889
#> 24                33.3642106
#> 25                32.5645185
#> 26                39.1181522
#> 27                42.7508229
#> 28                36.5665657
#> 29                41.6598053
#> 30                46.0792860
#> 31                40.8755141
#> 32                36.3943527
#> 33                36.7560848
#> 34                37.1618399
#> 35                45.7063943
#> 36                46.8911919
#> 37                33.8065282
#> 38                31.9940131
#> 39                35.9795648
#> 40                 0.1324663
#> 41                27.3183408
#> 42                39.8860660
#> 43                45.0329116
#> 44                32.8771479
#> 45                46.6570895
#> 46                32.9249894
#> 47                43.0704963
#> 48                30.6739362
#> 49                 1.0385094
#> 50                23.9370788
#> 51                32.2601590
#> 52                35.0776605
#> 53                40.9546553
#> 54                40.0329658
#> 55                33.7422656
#> 56                42.2319555
#> 57                42.4787733
#> 58                34.7329458
#> 59                40.3914451
#> 60                36.4035320
#> 61                52.6819670
#> 62                46.4150298
#> 63                43.1005115
#> 64                42.5504667
#> 65                32.1159709
#> 66                39.1408019
#> 67                34.2267966
#> 68                39.7774051
#> 69                35.6321313
#> 70                34.1539866
#> 71                36.8819717
#> 72                25.0736130
#> 73                33.3487750
#> 74                29.4171929
#> 75                48.6835092
#> 76                42.5471979
#> 77                32.2644305
#> 78                42.8430450
#> 79                38.6047672
#> 80                28.1289016
#> 81                 3.5430122
#> 82                42.0700875
#> 83                43.9588499
#> 84                44.5564816
#> 85                45.3402318
#> 86                30.1499099
#> 87                36.7924677
#> 88                42.5103666
#> 89                35.2071624
#> 90                43.4931602
#> 91                42.7810333
#> 92                39.5786253
#> 93                32.9940617
#> 94                27.3222974
#> 95                44.4942262
#> 96                76.1998546
#> 97                42.7338607
#> 98                36.1528516
#> 99                37.4844325
#> 100               43.0635237