Get confidence intervals for provider effects or standardized measures

Provide confidence intervals for provider effects or standardized measures from a fixed effect linear model.

Usage

# S3 method for linear_fe
confint(
  fit,
  parm,
  level = 0.95,
  option = "SM",
  stdz = "indirect",
  null = "median",
  alternative = "two.sided"
)

Arguments

fit

a model fitted from linear_fe.

parm

specify a subset of providers for which confidence intervals are given. By default, all providers are included. The class of parm should match the class of the provider IDs.

level

the confidence level. The default value is 0.95.

option

a character string specifying whether the confidence intervals should be provided for provider effects or standardized measures:

"gamma" provider effect (only supports "two.sided" confidence interval).

"SM" standardized measures.

stdz

a character string or a vector specifying the standardization method if option includes "SM". See stdz argument in SM_output.linear_fe.

null

a character string or a number specifying the population norm for calculating standardized measures if option includes "SM". See null argument in SM_output.linear_fe.

alternative

a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater", or "less". Note that "gamma" for argument option only supports "two.sided".

Value

A list of data frames containing the confidence intervals based on the values of option and stdz.

CI.gamma

Confidence intervals for provider effects if option includes "gamma".

CI.indirect

Confidence intervals for indirect standardized differences if option includes "SM" and stdz includes "indirect".

CI.direct

Confidence intervals for direct standardized differences if option includes "SM" and stdz includes "direct".

Examples

data(ExampleDataLinear)
outcome <- ExampleDataLinear$Y
covar <- ExampleDataLinear$Z
ID <- ExampleDataLinear$ID
fit_linear <- linear_fe(Y = outcome, Z = covar, ID = ID)
#> Input format: Y, Z, and ID.
confint(fit_linear)
#> $CI.indirect
#>     Indirect.Difference indirect.Lower indirect.Upper
#> 1          -0.370387492   -0.578496218    -0.16227877
#> 2          -0.864439821   -1.081249351    -0.64763029
#> 3          -0.707620305   -0.932825548    -0.48241506
#> 4           0.276001179    0.047772912     0.50422945
#> 5           0.263906784    0.048407294     0.47940627
#> 6           0.545099160    0.325596241     0.76460208
#> 7          -0.075159001   -0.303387269     0.15306927
#> 8          -0.000910955   -0.235569446     0.23374754
#> 9           0.034446109   -0.187853139     0.25674536
#> 10         -0.104258253   -0.326557501     0.11804099
#> 11         -0.446719918   -0.676506076    -0.21693376
#> 12          0.272124885    0.046919642     0.49733013
#> 13         -0.389098903   -0.608601822    -0.16959598
#> 14         -0.536111775   -0.781523500    -0.29070005
#> 15          0.239984666    0.017685419     0.46228391
#> 16          0.709655905    0.462304136     0.95700767
#> 17         -0.204927425   -0.441280223     0.03142537
#> 18         -0.559891638   -0.786593278    -0.33319000
#> 19          0.305127819    0.082828571     0.52742707
#> 20         -0.349044122   -0.563257037    -0.13483121
#> 21          0.539809597    0.305151106     0.77446809
#> 22         -0.066203547   -0.285706466     0.15329937
#> 23          0.170000547   -0.075411177     0.41541227
#> 24         -0.287521481   -0.508409291    -0.06663367
#> 25          0.615940315    0.405453126     0.82642750
#> 26         -0.319395397   -0.529882586    -0.10890821
#> 27          0.561550115    0.333321847     0.78977838
#> 28          0.503528347    0.293041158     0.71401554
#> 29         -0.075355715   -0.290855206     0.14014378
#> 30         -0.248823513   -0.463036428    -0.03461060
#> 31          0.073722417   -0.151482826     0.29892766
#> 32         -0.497946595   -0.717449514    -0.27844368
#> 33          0.089175592   -0.127633938     0.30598512
#> 34         -0.407584532   -0.628472342    -0.18669672
#> 35         -0.301273818   -0.527975458    -0.07457218
#> 36         -0.356371875   -0.569320986    -0.14342276
#> 37         -0.581289472   -0.790577294    -0.37200165
#> 38         -0.317345556   -0.547131714    -0.08755940
#> 39          0.372819229    0.164710503     0.58092796
#> 40         -0.965013385   -1.193241653    -0.73678512
#> 41          0.876132354    0.628780586     1.12348412
#> 42         -0.581068848   -0.836667929    -0.32546977
#> 43          0.309689704    0.091545949     0.52783346
#> 44          0.029360968   -0.178747758     0.23746969
#> 45         -0.125427582   -0.330115109     0.07925995
#> 46         -0.237415644   -0.464117284    -0.01071400
#> 47          0.252593359    0.034449604     0.47073711
#> 48          0.658136650    0.452327538     0.86394576
#> 49         -0.281625478   -0.493332893    -0.06991806
#> 50          0.171954742   -0.057831416     0.40174090
#> 51          0.402203905    0.197516377     0.60689143
#> 52         -0.435091528   -0.668091634    -0.20209142
#> 53          0.064973790   -0.158764303     0.28871188
#> 54          0.721207770    0.510720581     0.93169496
#> 55          0.045288463   -0.161660874     0.25223780
#> 56         -0.219211832   -0.453870323     0.01544666
#> 57          0.562919521    0.339181429     0.78665761
#> 58         -0.217467569   -0.414786020    -0.02014912
#> 59          0.479610640    0.249824482     0.70939680
#> 60          0.014263420   -0.231148305     0.25967514
#> 61         -0.341672988   -0.563972235    -0.11937374
#> 62          0.168812468   -0.047997062     0.38562200
#> 63          0.425351081    0.188998283     0.66170388
#> 64          0.112877281   -0.115350986     0.34110555
#> 65         -0.240767716   -0.468995984    -0.01253945
#> 66          0.113621679   -0.126232833     0.35347619
#> 67         -0.210017207   -0.438245475     0.01821106
#> 68         -0.377252972   -0.598140781    -0.15636516
#> 69          0.340108560    0.116370467     0.56384665
#> 70         -0.592022169   -0.823398562    -0.36064578
#> 71          0.359159317    0.148672128     0.56964651
#> 72         -0.577191699   -0.790140810    -0.36424259
#> 73          0.253202263    0.027997020     0.47840751
#> 74          0.161887778   -0.038490056     0.36226561
#> 75          0.329562851    0.125978768     0.53314693
#> 76         -0.474828215   -0.700033459    -0.24962297
#> 77          0.178203557   -0.088966888     0.44537400
#> 78          0.154776675   -0.048807408     0.35836076
#> 79         -0.059867540   -0.283605633     0.16387055
#> 80         -0.001693390   -0.223992638     0.22060586
#> 81          0.757613642    0.544664531     0.97056275
#> 82          0.117947073   -0.101555846     0.33744999
#> 83         -0.547038800   -0.776824958    -0.31725264
#> 84          0.601429101    0.395619989     0.80723821
#> 85          0.231636139    0.029137846     0.43413443
#> 86         -0.593113354   -0.809922884    -0.37630382
#> 87          0.413513827    0.203026638     0.62400102
#> 88          0.000910955   -0.224294288     0.22611620
#> 89          0.668264685    0.467886851     0.86864252
#> 90         -0.052249693   -0.267749183     0.16324980
#> 91         -0.346328614   -0.582681412    -0.10997582
#> 92         -1.183652486   -1.408857730    -0.95844724
#> 93          0.670704078    0.446965985     0.89444217
#> 94         -0.135793277   -0.343902003     0.07231545
#> 95         -0.011371410   -0.232259219     0.20951640
#> 96         -0.139075317   -0.348363138     0.07021251
#> 97          0.308817383    0.057443371     0.56019139
#> 98         -0.244303261   -0.450112373    -0.03849415
#> 99          0.242067535    0.005714736     0.47842033
#> 100        -0.301230370   -0.515443284    -0.08701745
#>