Get confidence intervals for provider effects or standardized measures from a fitted linear_re object

Provide confidence intervals for provider effects or standardized measures from a random effect linear model.

Usage

# S3 method for linear_re
confint(
  fit,
  parm,
  level = 0.95,
  option = "SM",
  stdz = "indirect",
  alternative = "two.sided"
)

Arguments

fit

a model fitted from linear_re.

parm

specify a subset of providers for which confidence intervals are given. By default, all providers are included. The class of parm should match the class of the provider IDs.

level

the confidence level. The default value is 0.95.

option

a character string specifying whether the confidence intervals should be provided for provider effects or standardized measures:

"alpha" provider effect.

"SM" standardized measures.

stdz

a character string or a vector specifying the standardization method if option includes "SM". See stdz argument in SM_output.linear_re.

alternative

a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater", or "less". Note that "alpha" for argument option only supports "two.sided".

Value

A list of data frames containing the confidence intervals based on the values of option and stdz.

CI.alpha

Confidence intervals for provider effects if option includes "alpha".

CI.indirect

Confidence intervals for indirect standardized differences if option includes "SM" and stdz includes "indirect".

CI.direct

Confidence intervals for direct standardized differences if option includes "SM" and stdz includes "direct".

Examples

data(ExampleDataLinear)
outcome <- ExampleDataLinear$Y
ID <- ExampleDataLinear$ID
covar <- ExampleDataLinear$Z
fit_re <- linear_re(Y = outcome, Z = covar, ID = ID)
#> Input format: Y, Z, and ID.
confint(fit_re)
#> $CI.indirect
#>     Indirect.Difference indirect.Lower indirect.Upper
#> 1          -0.338240958   -0.539449411   -0.137032504
#> 2          -0.794389223   -1.003428639   -0.585349807
#> 3          -0.646058872   -0.862592005   -0.429525739
#> 4           0.260222333    0.041006319    0.479438347
#> 5           0.253447824    0.045583295    0.461312354
#> 6           0.511178626    0.299728406    0.722628846
#> 7          -0.061265301   -0.280481315    0.157950712
#> 8           0.006948695   -0.217946528    0.231843917
#> 9           0.038982628   -0.174963777    0.252929033
#> 10         -0.089775701   -0.303722106    0.124170705
#> 11         -0.403966192   -0.624561591   -0.183370793
#> 12          0.258342952    0.041809819    0.474876085
#> 13         -0.353891953   -0.565342173   -0.142441733
#> 14         -0.480798587   -0.715106075   -0.246491098
#> 15          0.228403421    0.014457016    0.442349827
#> 16          0.652245404    0.416251313    0.888239495
#> 17         -0.180256778   -0.406642134    0.046128578
#> 18         -0.508231543   -0.726093728   -0.290369358
#> 19          0.288147391    0.074200986    0.502093796
#> 20         -0.317763248   -0.524472482   -0.111054015
#> 21          0.501834328    0.276939105    0.726729550
#> 22         -0.054292981   -0.265743201    0.157157239
#> 23          0.161368104   -0.072939384    0.395675593
#> 24         -0.258343295   -0.471030623   -0.045655968
#> 25          0.580695732    0.377340130    0.784051334
#> 26         -0.290963078   -0.494318680   -0.087607476
#> 27          0.524245442    0.305029429    0.743461456
#> 28          0.475868844    0.272513242    0.679224446
#> 29         -0.062803193   -0.270667722    0.145061337
#> 30         -0.224160405   -0.430869638   -0.017451171
#> 31          0.074646030   -0.141887103    0.291179163
#> 32         -0.454060915   -0.665511135   -0.242610695
#> 33          0.088741486   -0.120297930    0.297780902
#> 34         -0.369611317   -0.582298645   -0.156923990
#> 35         -0.270597627   -0.488459812   -0.052735442
#> 36         -0.324613417   -0.530186406   -0.119040428
#> 37         -0.534115124   -0.736388605   -0.331841643
#> 38         -0.285213112   -0.505808511   -0.064617714
#> 39          0.354251227    0.153042773    0.555459681
#> 40         -0.881389857   -1.100605871   -0.662173844
#> 41          0.803002465    0.567008374    1.038996556
#> 42         -0.517811990   -0.760936092   -0.274687887
#> 43          0.293492361    0.083257910    0.503726813
#> 44          0.035110130   -0.166098324    0.236318584
#> 45         -0.110060527   -0.308172084    0.088051029
#> 46         -0.211669804   -0.429531989    0.006192381
#> 47          0.241146710    0.030912258    0.451381162
#> 48          0.621420984    0.422293071    0.820548897
#> 49         -0.255612145   -0.460067423   -0.051156867
#> 50          0.165658248   -0.054937151    0.386253647
#> 51          0.383515095    0.185403539    0.581626651
#> 52         -0.392390843   -0.615824976   -0.168956710
#> 53          0.068123077   -0.147105035    0.283351189
#> 54          0.679024992    0.475669390    0.882380594
#> 55          0.048869088   -0.151290986    0.249029162
#> 56         -0.192475634   -0.417370857    0.032419588
#> 57          0.525940998    0.310712886    0.741169110
#> 58         -0.197817532   -0.389225270   -0.006409795
#> 59          0.448000079    0.227404680    0.668595478
#> 60          0.020624157   -0.213683332    0.254931645
#> 61         -0.308840599   -0.522787004   -0.094894194
#> 62          0.163885562   -0.045153854    0.372924977
#> 63          0.397151882    0.170766527    0.623537238
#> 64          0.110659002   -0.108557011    0.329875016
#> 65         -0.215252603   -0.434468617    0.003963410
#> 66          0.110843986   -0.118612712    0.340300685
#> 67         -0.185974771   -0.405190784    0.033241243
#> 68         -0.341368033   -0.554055360   -0.128680706
#> 69          0.320761519    0.105533408    0.535989631
#> 70         -0.537901540   -0.759902694   -0.315900385
#> 71          0.342325491    0.138969889    0.545681093
#> 72         -0.528870103   -0.734443092   -0.323297114
#> 73          0.240576715    0.024043582    0.457109848
#> 74          0.159376193   -0.034820204    0.353572590
#> 75          0.315175243    0.118064638    0.512285848
#> 76         -0.431149420   -0.647682554   -0.214616287
#> 77          0.166323929   -0.086693344    0.419341203
#> 78          0.152830117   -0.044280488    0.349940722
#> 79         -0.047475405   -0.262703516    0.167752707
#> 80          0.005401941   -0.208544464    0.219348346
#> 81          0.711346301    0.505773312    0.916919290
#> 82          0.115364610   -0.096085610    0.326814830
#> 83         -0.495816092   -0.716411491   -0.275220693
#> 84          0.569053793    0.369925880    0.768181706
#> 85          0.223492669    0.027367995    0.419617343
#> 86         -0.543695052   -0.752734468   -0.334655636
#> 87          0.393472922    0.190117320    0.596828524
#> 88          0.008107068   -0.208426065    0.224640202
#> 89          0.633800387    0.439603990    0.827996785
#> 90         -0.041475064   -0.249339594    0.166389465
#> 91         -0.310329631   -0.536714987   -0.083944276
#> 92         -1.084831958   -1.301365091   -0.868298824
#> 93          0.625010127    0.409782015    0.840238239
#> 94         -0.119090926   -0.320299380    0.082117528
#> 95         -0.004338301   -0.217025628    0.208349026
#> 96         -0.123299039   -0.325572521    0.078974442
#> 97          0.287366666    0.047887139    0.526846194
#> 98         -0.219943853   -0.419071766   -0.020815940
#> 99          0.229334049    0.002948693    0.455719404
#> 100        -0.272865484   -0.479574718   -0.066156250
#>