Get confidence intervals for provider effects or standardized measures from a fitted logis_fe object

Provide confidence intervals for provider effects or standardized measures from a fixed effect logistic model.

Usage

# S3 method for class 'logis_fe'
confint(
  object,
  parm,
  level = 0.95,
  test = "exact",
  option = "SM",
  stdz = "indirect",
  null = "median",
  measure = c("rate", "ratio"),
  alternative = "two.sided",
  ...
)

Arguments

object

a model fitted from logis_fe.

parm

specify a subset of providers for which confidence intervals are given. By default, all providers are included. The class of parm should match the class of the provider IDs.

level

the confidence level. The default value is 0.95.

test

a character string specifying the type of testing method. The default is "exact".

"exact" exact test.

"wald" wald test.

"score" score test.

option

a character string specifying whether the confidence intervals should be provided for provider effects or standardized measures:

"gamma" provider effect.

"SM" standardized measures.

stdz

a character string or a vector specifying the standardization method if option = "SM". See stdz argument in SM_output.logis_fe.

null

a character string or a number defining the population norm if option = "SM".

measure

a character string or a vector indicating whether the output measure is "ratio" or "rate" if option = "SM". Both "rate" and "ratio" will be provided by default.

"rate" output the standardized rate. The "rate" has been restricted to 0% - 100%.

"ratio" output the standardized ratio.

c("ratio", "rate") output both the standardized rate and ratio.

alternative

a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater", or "less". Note that "gamma" for argument option only supports "two.sided".

...

additional arguments that can be passed to the function.

Value

A dataframe (option = "gamma") or a list of data frames (option = "SM") containing the point estimate, and lower and upper bounds of the estimate.

Details

The wald test is invalid for extreme providers (i.e. when provider effect goes to infinity). We suggest using score or exact test to generate confidence intervals.

Examples

data(ExampleDataBinary)
outcome = ExampleDataBinary$Y
covar = ExampleDataBinary$Z
ProvID = ExampleDataBinary$ProvID
fit_fe <- logis_fe(Y = outcome, Z = covar, ProvID = ProvID, message = FALSE)
confint(fit_fe, option = "gamma")
#>           gamma gamma.lower  gamma.upper
#> 1   -1.32064745  -2.0218357 -0.627323497
#> 2   -0.66014873  -1.4662600  0.109920469
#> 3   -1.25964375  -1.8684628 -0.672249109
#> 4   -0.66541916  -1.2650808 -0.079973707
#> 5   -0.97153064  -1.6681618 -0.294288311
#> 6   -0.79873387  -1.4465916 -0.144441378
#> 7   -0.85806502  -1.4918434 -0.256731299
#> 8   -0.98061703  -1.6335029 -0.347265264
#> 9   -1.97198632  -2.6849658 -1.336489407
#> 10  -0.03380968  -0.7082366  0.643959756
#> 11  -0.40508001  -1.0240490  0.201374066
#> 12  -0.83731495  -1.4770807 -0.209565212
#> 13  -0.52427106  -1.2142344  0.187860911
#> 14  -1.41452378  -2.1197427 -0.751785448
#> 15  -0.84439047  -1.4826667 -0.226982274
#> 16  -1.63583137  -2.3419705 -0.981466766
#> 17  -0.37664227  -0.9775712  0.202150854
#> 18  -1.64669884  -2.3667000 -0.953864963
#> 19  -0.74999918  -1.2995353 -0.208854242
#> 20  -0.90503108  -1.4955244 -0.313595168
#> 21  -0.30957711  -0.9609071  0.356202734
#> 22   1.75760453   0.3567855  5.220978000
#> 23  -0.77650437  -1.4679457 -0.087524928
#> 24  -1.42410714  -2.0544219 -0.823645832
#> 25  -1.49611239  -2.1592634 -0.895031738
#> 26  -0.91441031  -1.6060302 -0.239253968
#> 27  -0.60505207  -1.2894393  0.079405194
#> 28  -1.13742216  -1.7554483 -0.538002158
#> 29  -0.69713348  -1.3496604 -0.061264004
#> 30  -0.32625987  -0.9210762  0.268225954
#> 31  -0.76213585  -1.4856135 -0.054216636
#> 32  -1.15179357  -1.7780178 -0.544965196
#> 33  -1.11822728  -1.8383948 -0.431970495
#> 34  -1.08384128  -1.7024360 -0.483012294
#> 35  -0.35719945  -1.0446115  0.309759252
#> 36  -0.26078465  -0.9746539  0.499695832
#> 37  -1.38084090  -2.1158850 -0.700191390
#> 38  -1.54487236  -2.3024156 -0.880234414
#> 39  -1.18770784  -1.8521123 -0.560997112
#> 40  -7.18906269        -Inf -4.348757371
#> 41  -1.99198648  -2.7995359 -1.258981850
#> 42  -0.84860235  -1.4891636 -0.228481096
#> 43  -0.41332497  -1.0826538  0.257835326
#> 44  -1.46698450  -2.1822765 -0.801468767
#> 45  -0.27819427  -0.8933000  0.345470308
#> 46  -1.46268548  -2.1442996 -0.808563251
#> 47  -0.57786467  -1.2259253  0.055440373
#> 48  -1.67188147  -2.3679489 -1.031368326
#> 49  -4.54884326        -Inf -2.287938760
#> 50  -2.34224073  -3.1915757 -1.594926149
#> 51  -1.52319624  -2.2539687 -0.843218253
#> 52  -1.26755091  -1.9804262 -0.590005600
#> 53  -0.75747754  -1.4299817 -0.095757883
#> 54  -0.83414670  -1.5814271 -0.108466549
#> 55  -1.38872235  -2.0317421 -0.776076819
#> 56  -0.64945066  -1.2262683 -0.085347867
#> 57  -0.62737237  -1.2873628 -0.007755105
#> 58  -1.29987873  -1.9422962 -0.684813729
#> 59  -0.80458347  -1.4558562 -0.169767482
#> 60  -1.15147120  -1.8115927 -0.515183844
#> 61   0.22188364  -0.4141957  0.920399121
#> 62  -0.29889400  -0.9560132  0.379528799
#> 63  -0.57454640  -1.3005555  0.122595855
#> 64  -0.62204106  -1.2960186  0.030469542
#> 65  -1.53584440  -2.3281545 -0.782727905
#> 66  -0.91317036  -1.5217853 -0.300528347
#> 67  -1.33821163  -2.2086846 -0.559777758
#> 68  -0.85710695  -1.5022663 -0.221005525
#> 69  -1.21887992  -1.8506722 -0.613391103
#> 70  -1.34972066  -2.0821473 -0.666125741
#> 71  -1.10902769  -1.8060845 -0.446733912
#> 72  -2.21471856  -3.1546226 -1.428393843
#> 73  -1.42510167  -2.0481670 -0.829856036
#> 74  -1.78984139  -2.5591776 -1.098348600
#> 75  -0.10922182  -0.6648861  0.464469576
#> 76  -0.62171345  -1.2698796  0.015216764
#> 77  -1.51990342  -2.2951469 -0.818338643
#> 78  -0.59723068  -1.2042810  0.012293213
#> 79  -0.95858328  -1.6299959 -0.314002568
#> 80  -1.91414214  -2.6890032 -1.206426864
#> 81  -2.49768668        -Inf -0.799989657
#> 82  -0.66308861  -1.2790730 -0.068453171
#> 83  -0.50316773  -1.0936030  0.112910877
#> 84  -0.45309702  -1.0701582  0.166782629
#> 85  -0.38777096  -1.0063152  0.227299130
#> 86  -1.71997793  -2.5165153 -0.983053666
#> 87  -1.11650406  -1.7522309 -0.502657520
#> 88  -0.62501624  -1.1765431 -0.064144424
#> 89  -1.25756420  -1.9282866 -0.616043102
#> 90  -0.54250260  -1.1947461  0.093469913
#> 91  -0.60255748  -1.2398876  0.039372701
#> 92  -0.87514468  -1.5195593 -0.226397868
#> 93  -1.45315456  -2.2094373 -0.785169365
#> 94  -1.99532199  -2.8277437 -1.201948156
#> 95  -0.45785978  -1.0274306  0.120100070
#> 96   2.06670990   0.7140124  5.508005148
#> 97  -0.60635817  -1.1851543 -0.027953488
#> 98  -1.16770373  -1.9660436 -0.433304819
#> 99  -1.05656024  -1.6670295 -0.469516002
#> 100 -0.57805028  -1.2032501  0.040339590
confint(fit_fe, option = "SM")
#> $CI.indirect_ratio
#>     indirect_ratio CI_ratio.lower CI_ratio.upper
#> 1        0.9023178      0.7402088      1.0672126
#> 2        1.1041807      0.7735631      1.4573133
#> 3        0.8983017      0.7355877      1.0694439
#> 4        1.0727439      0.8956600      1.2559052
#> 5        0.9760779      0.7411505      1.2207496
#> 6        1.0315284      0.8493057      1.2112588
#> 7        1.0164827      0.7888368      1.2619708
#> 8        0.9746555      0.7685617      1.1897858
#> 9        0.5956825      0.3983482      0.8284125
#> 10       1.2219800      1.0499583      1.3927180
#> 11       1.1899474      0.9580589      1.4287813
#> 12       1.0207006      0.8338553      1.2126256
#> 13       1.0884862      0.9294402      1.2406192
#> 14       0.8416191      0.6506159      1.0503876
#> 15       1.0193463      0.8204073      1.2280985
#> 16       0.7380075      0.5274581      0.9729249
#> 17       1.1877301      0.9769874      1.4100195
#> 18       0.7678055      0.5617378      0.9857338
#> 19       1.0512548      0.8753380      1.2309778
#> 20       1.0012013      0.8364918      1.1655371
#> 21       1.1399053      0.9879522      1.2871785
#> 22       1.1118533      1.0693340      1.1242796
#> 23       1.0294228      0.8763300      1.1833086
#> 24       0.8427782      0.6711483      1.0266900
#> 25       0.8014345      0.6242180      1.0050158
#> 26       0.9973722      0.8103211      1.1911267
#> 27       1.0895313      0.8895845      1.2895223
#> 28       0.9257552      0.7392608      1.1205644
#> 29       1.0587625      0.8788735      1.2453949
#> 30       1.1657115      0.9966398      1.3350412
#> 31       1.0313192      0.8772479      1.1900522
#> 32       0.9295431      0.7607372      1.1060687
#> 33       0.9447481      0.7754090      1.1259461
#> 34       0.9484637      0.7793304      1.1247396
#> 35       1.1724587      0.9586090      1.3956063
#> 36       1.1229547      0.9871250      1.2471375
#> 37       0.8182736      0.5813942      1.0825996
#> 38       0.7238168      0.4837468      1.0132470
#> 39       0.9097521      0.7220317      1.1137524
#> 40       0.0000000      0.0000000      0.2202662
#> 41       0.6490955      0.4446158      0.8834420
#> 42       1.0162138      0.8357135      1.2052477
#> 43       1.1286421      0.9551098      1.3014436
#> 44       0.7841368      0.5509559      1.0428597
#> 45       1.1938430      1.0048229      1.3792327
#> 46       0.8312215      0.6427554      1.0311354
#> 47       1.0998026      0.9065421      1.2994766
#> 48       0.7293409      0.5291971      0.9547792
#> 49       0.0000000      0.0000000      0.5156806
#> 50       0.5376522      0.3520556      0.7604584
#> 51       0.7745723      0.5488960      1.0249325
#> 52       0.8778571      0.6621564      1.1098373
#> 53       1.0490766      0.8334107      1.2698838
#> 54       1.0165289      0.8462982      1.1940120
#> 55       0.8529744      0.6771231      1.0414062
#> 56       1.0896157      0.8926921      1.2929312
#> 57       1.0858096      0.8883589      1.3015714
#> 58       0.8778500      0.6964388      1.0710168
#> 59       1.0286366      0.8488958      1.2148501
#> 60       0.9274987      0.7462065      1.1185915
#> 61       1.2235325      1.1025071      1.3293703
#> 62       1.1563515      0.9878586      1.3174462
#> 63       1.1060474      0.8788158      1.3468592
#> 64       1.0970078      0.8721913      1.3324784
#> 65       0.7959976      0.5676660      1.0417057
#> 66       0.9991318      0.8217254      1.1750436
#> 67       0.8336963      0.5714236      1.1416511
#> 68       1.0122806      0.8559383      1.1716972
#> 69       0.8978102      0.7092661      1.0979607
#> 70       0.8560419      0.6516148      1.0814627
#> 71       0.9273993      0.7046181      1.1672933
#> 72       0.4946818      0.2748953      0.7784619
#> 73       0.8310033      0.6478874      1.0263216
#> 74       0.6806241      0.4693187      0.9267015
#> 75       1.2527246      1.0771082      1.4206058
#> 76       1.0912934      0.8873032      1.3002798
#> 77       0.7756572      0.5506774      1.0349395
#> 78       1.0812262      0.9241213      1.2374942
#> 79       0.9826804      0.7872987      1.1897671
#> 80       0.6640426      0.4596421      0.8968505
#> 81       0.0000000      0.0000000      1.0962265
#> 82       1.0830968      0.8779316      1.2989710
#> 83       1.0978356      0.9549863      1.2323897
#> 84       1.1251051      0.9562863      1.2928238
#> 85       1.1529445      0.9718890      1.3353695
#> 86       0.7394821      0.5288972      0.9750521
#> 87       0.9306546      0.7360549      1.1351116
#> 88       1.0858777      0.9195885      1.2482474
#> 89       0.8907597      0.7007484      1.0933689
#> 90       1.1220075      0.9075022      1.3443720
#> 91       1.0755638      0.9186800      1.2309500
#> 92       1.0094794      0.8348974      1.1824343
#> 93       0.7886380      0.5685205      1.0511890
#> 94       0.7133298      0.5183287      0.9222288
#> 95       1.1216017      0.9679948      1.2720037
#> 96       1.1432459      1.0991945      1.1557293
#> 97       1.0849468      0.9231905      1.2468046
#> 98       0.9131695      0.6936897      1.1601938
#> 99       0.9513290      0.7690605      1.1439136
#> 100      1.1012385      0.9117197      1.2936342
#> 
#> $CI.indirect_rate
#>     indirect_rate CI_rate.lower CI_rate.upper
#> 1        34.98412      28.69893     41.377326
#> 2        42.81063      29.99213     56.502077
#> 3        34.82841      28.51977     41.463838
#> 4        41.59178      34.72599     48.693204
#> 5        37.84391      28.73544     47.330171
#> 6        39.99380      32.92877     46.962200
#> 7        39.41046      30.58431     48.928374
#> 8        37.78876      29.79821     46.129661
#> 9        23.09544      15.44452     32.118712
#> 10       47.37788      40.70835     53.997627
#> 11       46.13593      37.14529     55.395852
#> 12       39.57399      32.32974     47.015190
#> 13       42.20213      36.03570     48.100541
#> 14       32.63075      25.22529     40.724997
#> 15       39.52148      31.80834     47.615099
#> 16       28.61358      20.45029     37.721662
#> 17       46.04996      37.87917     54.668430
#> 18       29.76889      21.77936     38.218280
#> 19       40.75862      33.93808     47.726734
#> 20       38.81797      32.43196     45.189506
#> 21       44.19572      38.30429     49.905711
#> 22       43.10811      41.45958     43.589896
#> 23       39.91216      33.97654     45.878532
#> 24       32.67569      26.02136     39.806209
#> 25       31.07274      24.20181     38.965870
#> 26       38.66952      31.41728     46.181650
#> 27       42.24265      34.49044     49.996584
#> 28       35.89282      28.66218     43.445852
#> 29       41.04970      34.07516     48.285703
#> 30       45.19627      38.64112     51.761414
#> 31       39.98569      34.01213     46.139990
#> 32       36.03969      29.49484     42.883832
#> 33       36.62920      30.06369     43.654508
#> 34       36.77327      30.21573     43.607729
#> 35       45.45787      37.16661     54.109611
#> 36       43.53853      38.27222     48.353265
#> 37       31.72561      22.54147     41.973904
#> 38       28.06339      18.75554     39.285005
#> 39       35.27236      27.99418     43.181741
#> 40        0.00000       0.00000      8.540028
#> 41       25.16634      17.23838     34.252285
#> 42       39.40003      32.40178     46.729141
#> 43       43.75904      37.03095     50.458789
#> 44       30.40208      21.36133     40.433131
#> 45       46.28697      38.95839     53.474783
#> 46       32.22762      24.92053     39.978563
#> 47       42.64089      35.14791     50.382527
#> 48       28.27757      20.51771     37.018127
#> 49        0.00000       0.00000     19.993659
#> 50       20.84553      13.64969     29.484036
#> 51       30.03125      21.28146     39.738067
#> 52       34.03575      25.67273     43.029946
#> 53       40.67417      32.31250     49.235173
#> 54       39.41225      32.81217     46.293515
#> 55       33.07101      26.25301     40.376778
#> 56       42.24593      34.61092     50.128752
#> 57       42.09836      34.44292     50.463747
#> 58       34.03547      27.00191     41.524822
#> 59       39.88168      32.91288     47.101439
#> 60       35.96042      28.93147     43.369358
#> 61       47.43807      42.74574     51.541549
#> 62       44.83337      38.30066     51.079233
#> 63       42.88301      34.07292     52.219616
#> 64       42.53253      33.81608     51.662053
#> 65       30.86194      22.00921     40.388386
#> 66       38.73774      31.85944     45.558083
#> 67       32.32357      22.15489     44.263411
#> 68       39.24754      33.18593     45.428340
#> 69       34.80936      27.49924     42.569473
#> 70       33.18994      25.26402     41.929823
#> 71       35.95657      27.31903     45.257594
#> 72       19.17951      10.65808     30.182058
#> 73       32.21916      25.11950     39.791924
#> 74       26.38875      18.19614     35.929516
#> 75       48.56988      41.76099     55.078876
#> 76       42.31097      34.40199     50.413670
#> 77       30.07332      21.35054     40.126055
#> 78       41.92065      35.82948     47.979384
#> 79       38.09990      30.52467     46.128936
#> 80       25.74586      17.82097     34.772150
#> 81        0.00000       0.00000     42.502236
#> 82       41.99318      34.03864     50.362922
#> 83       42.56462      37.02616     47.781474
#> 84       43.62190      37.07656     50.124588
#> 85       44.70127      37.68150     51.774146
#> 86       28.67076      20.50608     37.804134
#> 87       36.08278      28.53788     44.009866
#> 88       42.10100      35.65373     48.396300
#> 89       34.53600      27.16900     42.391444
#> 90       43.50180      35.18513     52.123186
#> 91       41.70112      35.61851     47.725653
#> 92       39.13893      32.37014     45.844634
#> 93       30.57660      22.04234     40.756071
#> 94       27.65680      20.09633     35.756102
#> 95       43.48607      37.53051     49.317364
#> 96       44.32524      42.61731     44.809245
#> 97       42.06491      35.79339     48.340359
#> 98       35.40486      26.89532     44.982337
#> 99       36.88436      29.81755     44.351131
#> 100      42.69656      35.34865     50.156008
#>