Get confidence intervals for provider effects or standardized measures from a fitted logis_fe object

Provide confidence intervals for provider effects or standardized measures from a fixed effect logistic model.

Usage

# S3 method for logis_fe
confint(
  fit,
  parm,
  level = 0.95,
  test = "exact",
  option = "SM",
  stdz = "indirect",
  null = "median",
  measure = c("rate", "ratio"),
  alternative = "two.sided"
)

Arguments

fit

a model fitted from logis_fe.

parm

specify a subset of providers for which confidence intervals are given. By default, all providers are included. The class of parm should match the class of the provider IDs.

level

the confidence level. The default value is 0.95.

test

a character string specifying the type of testing method. The default is "exact".

"exact" exact test.

"wald" wald test.

"score" score test.

option

a character string specifying whether the confidence intervals should be provided for provider effects or standardized measures:

"gamma" provider effect.

"SM" standardized measures.

stdz

a character string or a vector specifying the standardization method if option = "SM". See stdz argument in SM_output.logis_fe.

null

a character string or a number defining the population norm if option = "SM".

measure

a character string or a vector indicating whether the output measure is "ratio" or "rate" if option = "SM". Both "rate" and "ratio" will be provided by default.

"rate" output the standardized rate. The "rate" has been restricted to 0% - 100%.

"ratio" output the standardized ratio.

c("ratio", "rate") output both the standardized rate and ratio.

alternative

a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater", or "less". Note that "gamma" for argument option only supports "two.sided".

Value

A dataframe (option = "gamma") or a list of data frames (option = "SM") containing the point estimate, and lower and upper bounds of the estimate.

Details

The wald test is invalid for extreme providers (i.e. when provider effect goes to infinity). We suggest using score or exact test to generate confidence intervals.

Examples

data(ExampleDataBinary)
outcome = ExampleDataBinary$Y
covar = ExampleDataBinary$Z
ID = ExampleDataBinary$ID
fit_fe <- logis_fe(Y = outcome, Z = covar, ID = ID, message = FALSE)
confint(fit_fe, option = "gamma")
#>            gamma gamma.lower  gamma.upper
#> 1    -1.34797989  -2.0537002 -0.646004794
#> 2    -0.67625337  -1.4783625  0.108280207
#> 3    -1.28525031  -1.8934880 -0.687370098
#> 4    -0.67938233  -1.2797033 -0.086092562
#> 5    -0.99265740  -1.6883067 -0.306586937
#> 6    -0.81532974  -1.4691758 -0.158365019
#> 7    -0.87543592  -1.5047498 -0.261985376
#> 8    -1.00068080  -1.6532438 -0.357317086
#> 9    -2.01487404  -2.7118054 -1.356036731
#> 10   -0.03521049  -0.7151699  0.646111807
#> 11   -0.41361385  -1.0331518  0.199699393
#> 12   -0.85510502  -1.4958132 -0.220165172
#> 13   -0.53422603  -1.2356102  0.178346095
#> 14   -1.44522997  -2.1448296 -0.766708116
#> 15   -0.86171651  -1.4996089 -0.234152168
#> 16   -1.67040261  -2.3682618 -0.997904352
#> 17   -0.38402254  -0.9825887  0.203794375
#> 18   -1.68115574  -2.3994015 -0.975923171
#> 19   -0.76531946  -1.3164409 -0.218223501
#> 20   -0.92383165  -1.5190103 -0.328029770
#> 21   -0.31598883  -0.9757455  0.350997995
#> 22    2.10966982   0.3349658  5.209107007
#> 23   -0.79288236  -1.4900974 -0.096824052
#> 24   -1.45276650  -2.0803594 -0.839562234
#> 25   -1.52753761  -2.1798670 -0.905815116
#> 26   -0.93347260  -1.6273596 -0.247829872
#> 27   -0.61772976  -1.3071059  0.071794357
#> 28   -1.16019294  -1.7779386 -0.551318825
#> 29   -0.71185643  -1.3658122 -0.066162358
#> 30   -0.33318626  -0.9325393  0.265798379
#> 31   -0.77986366  -1.5054646 -0.062132886
#> 32   -1.17567914  -1.8019958 -0.558866645
#> 33   -1.14317960  -1.8612181 -0.441887844
#> 34   -1.10685453  -1.7255004 -0.497054091
#> 35   -0.36491386  -1.0532923  0.313468992
#> 36   -0.26419627  -0.9965540  0.491294276
#> 37   -1.41172803  -2.1378951 -0.712369181
#> 38   -1.58135752  -2.3192928 -0.889554893
#> 39   -1.21308490  -1.8726683 -0.572213081
#> 40  -10.71003595        -Inf -4.449114094
#> 41   -2.04040784  -2.8338347 -1.284182751
#> 42   -0.86612613  -1.5067837 -0.235425596
#> 43   -0.42229711  -1.0979568  0.254360346
#> 44   -1.49821614  -2.2055709 -0.815138939
#> 45   -0.28407706  -0.9057786  0.341775719
#> 46   -1.49373950  -2.1736451 -0.827343803
#> 47   -0.59024463  -1.2395408  0.051745937
#> 48   -1.70754053  -2.3941624 -1.048207588
#> 49   -8.08483094        -Inf -2.327590188
#> 50   -2.39811515  -3.2284479 -1.618721988
#> 51   -1.55619892  -2.2789772 -0.858229080
#> 52   -1.29496181  -2.0038880 -0.603276054
#> 53   -0.77298706  -1.4471202 -0.103876054
#> 54   -0.85328390  -1.6009438 -0.116481623
#> 55   -1.41767677  -2.0574086 -0.792812684
#> 56   -0.66242978  -1.2398346 -0.091294432
#> 57   -0.64115255  -1.2957887 -0.006435919
#> 58   -1.32646278  -1.9666843 -0.699596625
#> 59   -0.82199750  -1.4738207 -0.178134425
#> 60   -1.17485306  -1.8346318 -0.526685839
#> 61    0.22681516  -0.4295697  0.913826264
#> 62   -0.30464309  -0.9726764  0.373924824
#> 63   -0.58766599  -1.3127715  0.122904080
#> 64   -0.63497696  -1.3084996  0.028114194
#> 65   -1.56981768  -2.3580068 -0.800869968
#> 66   -0.93148088  -1.5457470 -0.315060646
#> 67   -1.37293507  -2.2276349 -0.564528600
#> 68   -0.87563371  -1.5238313 -0.231888314
#> 69   -1.24453551  -1.8737814 -0.628077460
#> 70   -1.37964221  -2.1055889 -0.677810694
#> 71   -1.13189518  -1.8252968 -0.455360436
#> 72   -2.27495244  -3.1825218 -1.444903513
#> 73   -1.45420261  -2.0742817 -0.847440975
#> 74   -1.83001701  -2.5858310 -1.112767866
#> 75   -0.11220514  -0.6757168  0.460139170
#> 76   -0.63525849  -1.2852645  0.009311440
#> 77   -1.55579587  -2.3180779 -0.830171360
#> 78   -0.60979649  -1.2215577  0.003150925
#> 79   -0.97870858  -1.6492041 -0.321498826
#> 80   -1.95822469  -2.7208826 -1.228884058
#> 81   -6.23331119        -Inf -0.768916759
#> 82   -0.67639817  -1.2911662 -0.072201714
#> 83   -0.51406360  -1.1155439  0.100016951
#> 84   -0.46296110  -1.0851628  0.160617546
#> 85   -0.39609871  -1.0184524  0.224715876
#> 86   -1.75831645  -2.5468001 -0.999400891
#> 87   -1.13991691  -1.7742716 -0.516148679
#> 88   -0.63843087  -1.1958845 -0.076366251
#> 89   -1.28315787  -1.9520393 -0.628462126
#> 90   -0.55396407  -1.2064798  0.090692052
#> 91   -0.61513058  -1.2592403  0.031179117
#> 92   -0.89304494  -1.5437139 -0.239978505
#> 93   -1.48728104  -2.2268665 -0.791843698
#> 94   -2.04000365  -2.8684329 -1.230668884
#> 95   -0.46827940  -1.0447749  0.112325167
#> 96    2.43108467   0.6994034  5.498454728
#> 97   -0.61918928  -1.2021864 -0.036281584
#> 98   -1.19743934  -1.9854846 -0.441504316
#> 99   -1.07806889  -1.6874791 -0.480059312
#> 100  -0.59084370  -1.2185063  0.033368998
confint(fit_fe, option = "SR")
#> Error: Argument 'option' NOT as required!