Get confidence intervals for provider effects or standardized Mmeasures

Provide confidence intervals for provider effects or standardized measures from a fixed effect logistic model.

Usage

# S3 method for logis_fe
confint(
  fit,
  parm,
  level = 0.95,
  test = "exact",
  option = "SM",
  stdz = "indirect",
  null = "median",
  measure = c("rate", "ratio"),
  alternative = "two.sided"
)

Arguments

fit

a model fitted from logis_fe.

parm

specify a subset of providers for which confidence intervals are given. By default, all providers are included. The class of parm should match the class of the provider IDs.

level

the confidence level. The default value is 0.95.

test

a character string specifying the type of testing method. The default is "exact".

"exact" exact test.

"wald" wald test.

"score" score test.

option

a character string specifying whether the confidence intervals should be provided for provider effects or standardized measures:

"gamma" provider effect.

"SM" standardized measures.

stdz

a character string or a vector specifying the standardization method if option = "SM". See stdz argument in SM_output.logis_fe.

null

a character string or a number defining the population norm if option = "SM".

measure

a character string or a vector indicating whether the output measure is "ratio" or "rate" if option = "SM". Both "rate" and "ratio" will be provided by default.

"rate" output the standardized rate. The "rate" has been restricted to 0% - 100%.

"ratio" output the standardized ratio.

c("ratio", "rate") output both the standardized rate and ratio.

alternative

a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater", or "less".

Value

A dataframe (option = "gamma") or a list of data frames (option = "SM") containing the point estimate, and lower and upper bounds of the estimate.

Details

The wald test is invalid for extreme providers (i.e. when provider effect goes to infinity). We suggest using score or exact test to generate confidence intervals.

Examples

data(ExampleDataBinary)
outcome = ExampleDataBinary$Y
covar = ExampleDataBinary$Z
ID = ExampleDataBinary$ID

fit_fe <- logis_fe(Y = outcome, Z = covar, ID = ID, message = FALSE)
confint(fit_fe, option = "gamma")
#>           gamma gamma.lower   gamma.upper
#> 1   -1.23906651  -1.8874535 -0.5919709454
#> 2   -0.25264377  -0.8546339  0.3550735361
#> 3   -1.00234879  -1.6644820 -0.3407840961
#> 4   -2.60116916  -3.3955214 -1.8661539693
#> 5   -1.83949843  -2.5375004 -1.1626717141
#> 6   -0.44702906  -1.0413625  0.1629616772
#> 7   -1.57330339  -2.2927559 -0.8464769195
#> 8   -0.60159260  -1.2265142  0.0245516879
#> 9   -0.90765758  -1.5225940 -0.3002451620
#> 10  -1.40569110  -2.0651922 -0.7795192744
#> 11  -1.02751018  -1.6963661 -0.3835260841
#> 12  -0.86138571  -1.5138256 -0.2163765773
#> 13  -1.15874493  -1.7733808 -0.5490585192
#> 14  -1.58604625  -2.2819711 -0.9158909981
#> 15  -1.33758218  -2.0717338 -0.6402988166
#> 16  -0.25452003  -0.8395087  0.3357237078
#> 17  -0.39394949  -1.0421874  0.2603587184
#> 18  -0.85664058  -1.4329171 -0.2867701789
#> 19  -0.70800492  -1.3632916 -0.0699854367
#> 20  -1.52870253  -2.2332768 -0.8521539447
#> 21  -0.37902569  -1.0207016  0.2633774211
#> 22  -1.33191243  -1.9614705 -0.7206903688
#> 23  -1.04462369  -1.6374382 -0.4633039238
#> 24  -1.17790702  -1.8300744 -0.5295786110
#> 25  -2.15029399  -2.8920137 -1.4526200157
#> 26  -0.95919499  -1.5637709 -0.3637756151
#> 27  -0.20424096  -0.7919753  0.3866960121
#> 28  -1.23244563  -1.9651710 -0.5324224025
#> 29   6.98694167   1.2135367           Inf
#> 30  -2.09702791  -2.9500310 -1.3095378883
#> 31  -1.34951405  -2.0116624 -0.7080876242
#> 32  -0.93601268  -1.5129706 -0.3644246217
#> 33  -0.71432753  -1.3091337 -0.1172504753
#> 34   0.05375552  -0.5841663  0.7083559131
#> 35   0.18352387  -0.4342867  0.8186391697
#> 36  -1.59649962  -2.2888082 -0.9344070419
#> 37  -0.33641237  -0.9873692  0.3098911099
#> 38  -0.05817828  -0.6291123  0.5241872396
#> 39  -0.55772674  -1.1595243  0.0486617179
#> 40  -0.95794451  -1.6157724 -0.3171006281
#> 41  -5.43108003        -Inf -0.2717240600
#> 42  -1.47418036  -2.1594183 -0.8290381215
#> 43  -0.79597911  -1.3776240 -0.2171739251
#> 44  -1.19027322  -1.7893221 -0.6035016068
#> 45  -1.00182628  -1.6409210 -0.3763220177
#> 46  -1.41751194  -2.1260693 -0.7294340892
#> 47  -1.06085509  -1.6950040 -0.4470088944
#> 48  -1.83780613  -2.5036200 -1.1936836639
#> 49  -1.05506565  -1.6711000 -0.4543562595
#> 50  -0.41319766  -1.0724015  0.2447697471
#> 51  -1.08930395  -1.7573358 -0.4341633599
#> 52  -0.86191082  -1.4861117 -0.2331233926
#> 53  -0.63072094  -1.2550995  0.0008036938
#> 54  -1.09858046  -1.7797390 -0.4314555148
#> 55  -0.42610008  -1.0572562  0.2049938619
#> 56  -0.87606554  -1.5031861 -0.2584024264
#> 57  -1.61225884  -2.2763839 -0.9671300256
#> 58  -1.39961474  -2.0851807 -0.7422622347
#> 59  -0.45819763  -1.0879243  0.1785277911
#> 60  -1.10297630  -1.6736400 -0.5416302582
#> 61  -1.56515734  -2.2262515 -0.9314688348
#> 62  -0.92826647  -1.5435541 -0.3148576217
#> 63  -1.31965662  -1.9477911 -0.7051120027
#> 64  -5.53745708        -Inf -0.0515573878
#> 65  -1.25661147  -1.8643661 -0.6544265993
#> 66  -1.44553435  -2.1132177 -0.8022469501
#> 67  -1.04104258  -1.6792497 -0.4190718986
#> 68  -0.93165681  -1.5580939 -0.3103176103
#> 69  -1.65025773  -2.3808504 -0.9472971670
#> 70  -1.23685046  -1.8658119 -0.6183160657
#> 71  -0.50583550  -1.0971809  0.0818293931
#> 72  -1.17210385  -1.8589650 -0.4981755181
#> 73  -0.77420239  -1.4352778 -0.1228039081
#> 74  -0.22223604  -0.7427706  0.3039437959
#> 75  -1.13812653  -1.8882929 -0.3995377266
#> 76  -0.98279107  -1.6110480 -0.3645160390
#> 77  -1.10790564  -1.7600469 -0.4735885181
#> 78  -0.60473123  -1.2490667  0.0344928596
#> 79  -1.15225863  -1.7925728 -0.5154042356
#> 80  -1.38927533  -2.1033556 -0.7001654748
#> 81  -1.31872130  -1.9327019 -0.7203417968
#> 82   0.09946910  -0.5600237  0.7725015976
#> 83  -1.28891810  -2.0550396 -0.5340156660
#> 84  -0.91418068  -1.5399765 -0.2861307301
#> 85  -0.54600072  -1.2936530  0.2067937758
#> 86  -1.75330947  -2.4256166 -1.1083667129
#> 87  -1.09379958  -1.7987195 -0.3836992876
#> 88  -0.82260731  -1.4960238 -0.1640491692
#> 89  -1.96616674  -2.6974502 -1.2740237919
#> 90  -0.41969680  -1.1065341  0.2467255443
#> 91  -0.46704655  -1.1167074  0.1783302285
#> 92  -0.32624268  -0.9078263  0.2662358697
#> 93   7.55928602   1.9280495           Inf
#> 94  -1.56408102  -2.2117702 -0.9214881670
#> 95   0.78098713   0.1151690  1.4623986627
#> 96  -1.24584080  -1.8385598 -0.6683219849
#> 97  -0.98488890  -1.7007537 -0.2861878935
#> 98  -6.20430717        -Inf -0.7393787322
#> 99  -1.16540786  -1.7780652 -0.5646894548
#> 100 -1.19815271  -1.8458817 -0.5545484521
confint(fit_fe, option = "SR")
#> Error: Argument 'option' NOT as required!