Result Summaries of Covariate Estimates from a fitted linear_fe or linear_re object

Provide the summary statistics for the covariate estimates for a fixed/random effect linear model.

Usage

# S3 method for linear_fe
summary(fit, parm, level = 0.95, null = 0, alternative = "two.sided")

# S3 method for linear_re
summary(fit, parm, level = 0.95, null = 0, alternative = "two.sided")

Arguments

fit

a model fitted from linear_fe or linear_re.

parm

Specifies a subset of covariates for which the result summaries should be output. By default, all covariates are included.

level

the confidence level during the hypothesis test, meaning a significance level of \(1 - \text{level}\). The default value is 0.95.

null

a number defining the null hypothesis for the covariate estimates. The default value is 0.

alternative

a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater", or "less".

Value

A data frame containing summary statistics for covariate estimates, with the following columns:

Estimate

the estimates of covariate coefficients.

Std.Error

the standard error of the estimate.

Stat

the test statistic.

p value

the p-value for the hypothesis test.

CI.upper

the lower bound of the confidence interval.

CI.lower

the upper bound of the confidence interval.

Examples

data(ExampleDataLinear)
outcome <- ExampleDataLinear$Y
covar <- ExampleDataLinear$Z
ID <- ExampleDataLinear$ID
fit_fe <- linear_fe(Y = outcome, Z = covar, ID = ID)
#> Input format: Y, Z, and ID.
summary(fit_fe)
#>     Estimate  Std.Error     Stat      p value  CI.Lower CI.Upper
#> z1 1.0019028 0.01160106 86.36305 < 2.2204e-16 0.9791651 1.024640
#> z2 1.0063310 0.01133705 88.76478 < 2.2204e-16 0.9841108 1.028551
#> z3 0.9845622 0.01158151 85.01157 < 2.2204e-16 0.9618629 1.007262
#> z4 0.9903556 0.01140266 86.85303 < 2.2204e-16 0.9680068 1.012704
#> z5 1.0071783 0.01160058 86.82140 < 2.2204e-16 0.9844416 1.029915

data(ExampleDataLinear)
outcome <- ExampleDataLinear$Y
covar <- ExampleDataLinear$Z
ID <- ExampleDataLinear$ID
fit_re <- linear_fe(Y = outcome, Z = covar, ID = ID)
#> Input format: Y, Z, and ID.
summary(fit_re)
#>     Estimate  Std.Error     Stat      p value  CI.Lower CI.Upper
#> z1 1.0019028 0.01160106 86.36305 < 2.2204e-16 0.9791651 1.024640
#> z2 1.0063310 0.01133705 88.76478 < 2.2204e-16 0.9841108 1.028551
#> z3 0.9845622 0.01158151 85.01157 < 2.2204e-16 0.9618629 1.007262
#> z4 0.9903556 0.01140266 86.85303 < 2.2204e-16 0.9680068 1.012704
#> z5 1.0071783 0.01160058 86.82140 < 2.2204e-16 0.9844416 1.029915