Result Summaries of Covariate Estimates from a fitted logis_fe object

Provide the summary statistics for the covariate estimates for a fixed effect logistic model.

Usage

# S3 method for logis_fe
summary(
  object,
  parm,
  level = 0.95,
  test = "wald",
  null = 0,
  alternative = "two.sided",
  ...
)

Arguments

object

a model fitted from logis_fe.

parm

Specifies a subset of covariates for which the result summaries should be output. By default, all covariates are included.

level

the confidence level during the hypothesis test, meaning a significance level of \(1 - \text{level}\). The default value is 0.95.

test

a character string specifying the type of testing method. The default is "wald".

"wald"wald test.

"lr"likelihood ratio test.

"score"score test.

null

a number defining the null hypothesis for the covariate estimates. The default value is 0.

alternative

a character string specifying the alternative hypothesis when test = "wald", must be one of "two.sided" (default), "greater", or "less".

...

additional arguments that can be passed to the function.

Value

A data frame containing summary statistics for covariate estimates, with the following columns:

Estimate

the estimates of covariate coefficients.

Std.Error

the standard error of the estimate, included only when test = "wald".

Stat

the test statistic.

p value

the p-value for the hypothesis test.

CI.upper

the lower bound of the confidence interval, included only when test = "wald".

CI.lower

the upper bound of the confidence interval, included only when test = "wald".

Examples

data(ExampleDataBinary)
outcome = ExampleDataBinary$Y
covar = ExampleDataBinary$Z
ID = ExampleDataBinary$ID

fit_fe <- logis_fe(Y = outcome, Z = covar, ID = ID, message = FALSE)
summary.wald <- summary(fit_fe, level = 0.95, test = "wald")
summary.wald
#>     Estimate  Std.Error     Stat p value  CI.Lower CI.Upper
#> z1 1.0512744 0.04179173 25.15508       0 0.9693641 1.133185
#> z2 1.0229572 0.04093213 24.99154       0 0.9427316 1.103183
#> z3 0.9918962 0.04035333 24.58028       0 0.9128051 1.070987
#> z4 1.0413031 0.04115101 25.30443       0 0.9606486 1.121958
#> z5 1.0930890 0.04173967 26.18825       0 1.0112808 1.174897