Conduct hypothesis testing for provider effects from a fitted linear_re object

Conduct hypothesis tests on provider effects and identify outlying providers for a random effect linear model.

Usage

# S3 method for linear_re
test(fit, parm, level = 0.95, null = 0, alternative = "two.sided", ...)

Arguments

fit

a model fitted from linear_re.

parm

specifies a subset of providers for which confidence intervals are to be given. By default, all providers are included. The class of parm should match the class of the provider IDs.

level

the confidence level during the hypothesis test, meaning a significance level of \(1 - \text{level}\). The default value is 0.95.

null

a number defining the null hypothesis for the provider effects. The default value is 0.

alternative

a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater", or "less".

...

additional arguments that can be passed to the function.

Value

A data frame containing the results of the hypothesis test, with the following columns:

flag

a flagging indicator where 1 means statistically higher than expected and -1 means statistically lower than expected.

p-value

the p-value of the hypothesis test.

stat

the test statistic.

Std.Error

the standard error of the provider effect estimate.

Details

The function identifies outlying providers based on hypothesis test results. For two-sided tests, 1 indicates performance significantly higher than expected, -1 indicates lower, For one-sided tests, 1 (right-tailed) or -1 (left-tailed) flags are used. Providers whose performance falls within the central range are flagged as 0. Outlying providers are determined by the test statistic falling beyond the threshold based on the significance level \(1 - \text{level}\).

Examples

data(ExampleDataLinear)
outcome <- ExampleDataLinear$Y
ID <- ExampleDataLinear$ID
covar <- ExampleDataLinear$Z
fit_re <- linear_re(Y = outcome, Z = covar, ID = ID)
#> Input format: Y, Z, and ID.
test(fit_re)
#>     flag      p value        stat  Std.Error
#> 1     -1 9.849440e-04 -3.29479245 0.10265926
#> 2     -1 9.459100e-14 -7.44823296 0.10665472
#> 3     -1 4.979833e-09 -5.84784463 0.11047812
#> 4      1 1.998695e-02  2.32659281 0.11184696
#> 5      1 1.685888e-02  2.38977092 0.10605528
#> 6      1 2.156337e-06  4.73819179 0.10788475
#> 7      0 5.838566e-01 -0.54776009 0.11184696
#> 8      0 9.517113e-01  0.06055794 0.11474457
#> 9      0 7.210019e-01  0.35712003 0.10915833
#> 10     0 4.108290e-01 -0.82243560 0.10915833
#> 11    -1 3.317048e-04 -3.58919176 0.11255074
#> 12     1 1.936608e-02  2.33840833 0.11047812
#> 13    -1 1.037049e-03 -3.28027789 0.10788475
#> 14    -1 5.774456e-05 -4.02184292 0.11954683
#> 15     1 3.640232e-02  2.09240477 0.10915833
#> 16     1 6.061092e-08  5.41698945 0.12040736
#> 17     0 1.186184e-01 -1.56059915 0.11550485
#> 18    -1 4.825661e-06 -4.57222771 0.11115622
#> 19     1 8.297466e-03  2.63971955 0.10915833
#> 20    -1 2.587218e-03 -3.01294969 0.10546583
#> 21     1 1.222753e-05  4.37349090 0.11474457
#> 22     0 6.147886e-01 -0.50324983 0.10788475
#> 23     0 1.770700e-01  1.34983169 0.11954683
#> 24    -1 1.728004e-02 -2.38069452 0.10851594
#> 25     1 2.183314e-08  5.59681027 0.10375476
#> 26    -1 5.042055e-03 -2.80433462 0.10375476
#> 27     1 2.770122e-06  4.68716755 0.11184696
#> 28     1 4.507880e-06  4.58647702 0.10375476
#> 29     0 5.537340e-01 -0.59217412 0.10605528
#> 30    -1 3.355062e-02 -2.12543152 0.10546583
#> 31     0 4.992543e-01  0.67566348 0.11047812
#> 32    -1 2.567770e-05 -4.20875911 0.10788475
#> 33     0 4.053838e-01  0.83204460 0.10665472
#> 34    -1 6.590875e-04 -3.40605564 0.10851594
#> 35    -1 1.491690e-02 -2.43439036 0.11115622
#> 36    -1 1.968703e-03 -3.09491343 0.10488611
#> 37    -1 2.274218e-07 -5.17540114 0.10320265
#> 38    -1 1.127415e-02 -2.53408471 0.11255074
#> 39     1 5.590355e-04  3.45074788 0.10265926
#> 40    -1 3.330669e-15 -7.88032018 0.11184696
#> 41     1 2.574681e-11  6.66904796 0.12040736
#> 42    -1 2.987962e-05 -4.17438190 0.12404519
#> 43     1 6.216137e-03  2.73615696 0.10726445
#> 44     0 7.323460e-01  0.34200646 0.10265926
#> 45     0 2.762180e-01 -1.08885455 0.10107918
#> 46     0 5.687696e-02 -1.90425517 0.11115622
#> 47     1 2.456654e-02  2.24815135 0.10726445
#> 48     1 9.566232e-10  6.11648427 0.10159774
#> 49    -1 1.427104e-02 -2.45036765 0.10431583
#> 50     0 1.410603e-01  1.47185391 0.11255074
#> 51     1 1.481175e-04  3.79420458 0.10107918
#> 52    -1 5.773191e-04 -3.44205208 0.11399910
#> 53     0 5.350212e-01  0.62035938 0.10981228
#> 54     1 5.968732e-11  6.54451864 0.10375476
#> 55     0 6.322764e-01  0.47852526 0.10212436
#> 56     0 9.345902e-02 -1.67742697 0.11474457
#> 57     1 1.672347e-06  4.78945527 0.10981228
#> 58    -1 4.280594e-02 -2.02559857 0.09765880
#> 59     1 6.879141e-05  3.98042763 0.11255074
#> 60     0 8.630292e-01  0.17251947 0.11954683
#> 61    -1 4.665139e-03 -2.82929012 0.10915833
#> 62     0 1.243915e-01  1.53659920 0.10665472
#> 63     1 5.851628e-04  3.43839991 0.11550485
#> 64     0 3.224779e-01  0.98937872 0.11184696
#> 65     0 5.428844e-02 -1.92452797 0.11184696
#> 66     0 3.437393e-01  0.94680270 0.11707190
#> 67     0 9.636027e-02 -1.66276107 0.11184696
#> 68    -1 1.656405e-03 -3.14578710 0.10851594
#> 69     1 3.489113e-03  2.92099866 0.10981228
#> 70    -1 2.044971e-06 -4.74892865 0.11326798
#> 71     1 9.690166e-04  3.29937128 0.10375476
#> 72    -1 4.599031e-07 -5.04232759 0.10488611
#> 73     1 2.943612e-02  2.17759606 0.11047812
#> 74     0 1.077182e-01  1.60853447 0.09908162
#> 75     1 1.724782e-03  3.13393652 0.10056848
#> 76    -1 9.517375e-05 -3.90257751 0.11047812
#> 77     0 1.976048e-01  1.28840576 0.12909282
#> 78     0 1.285959e-01  1.51966215 0.10056848
#> 79     0 6.654999e-01 -0.43233239 0.10981228
#> 80     0 9.605310e-01  0.04948721 0.10915833
#> 81     1 1.184552e-11  6.78208327 0.10488611
#> 82     0 2.849201e-01  1.06933197 0.10788475
#> 83    -1 1.056535e-05 -4.40526724 0.11255074
#> 84     1 2.130601e-08  5.60104771 0.10159774
#> 85     1 2.551830e-02  2.23346493 0.10006545
#> 86    -1 3.437835e-07 -5.09771191 0.10665472
#> 87     1 1.492368e-04  3.79233593 0.10375476
#> 88     0 9.415024e-01  0.07338167 0.11047812
#> 89     1 1.587182e-10  6.39675067 0.09908162
#> 90     0 6.957453e-01 -0.39107025 0.10605528
#> 91    -1 7.215661e-03 -2.68672370 0.11550485
#> 92    -1 0.000000e+00 -9.81942825 0.11047812
#> 93     1 1.258372e-08  5.69162331 0.10981228
#> 94     0 2.460243e-01 -1.16006023 0.10265926
#> 95     0 9.681103e-01 -0.03997847 0.10851594
#> 96     0 2.321935e-01 -1.19472743 0.10320265
#> 97     1 1.867855e-02  2.35188503 0.12218568
#> 98    -1 3.039919e-02 -2.16484984 0.10159774
#> 99     1 4.708968e-02  1.98549272 0.11550485
#> 100   -1 9.674801e-03 -2.58724060 0.10546583