Conduct hypothesis testing for provider effects from a fitted logis_fe object

Conduct hypothesis tests on provider effects and identify outlying providers for a fixed effect logistic model.

Usage

# S3 method for logis_fe
test(
  fit,
  parm,
  level = 0.95,
  test = "exact.poisbinom",
  score_modified = TRUE,
  null = "median",
  n = 10000,
  threads = 1,
  alternative = "two.sided",
  ...
)

Arguments

fit

a model fitted from logis_fe.

parm

specifies a subset of providers for which confidence intervals are to be given. By default, all providers are included. The class of parm should match the class of the provider IDs.

level

the confidence level during the hypothesis test, meaning a significance level of \(1 - \text{level}\). The default value is 0.95.

test

a character string specifying the type of testing method to be conducted. The default is "exact.poisbinom".

"exact.poisbinom"exact test based on Poisson-binomial distribution of \(O_i|Z_i\).

"exact.bootstrap"exact test based on bootstrap procedure.

"wald"wald test.

"score"score test.

score_modified

a logical indicating whether to use the modified score test ignoring the randomness of covariate coefficient for score teat ("test = score"). The default value is TRUE.

null

a character string or a number specifying null hypotheses of fixed provider effects. The default is "median".

n

resample size for bootstrapping when ("test = exact.bootstrap"). The default value is 10,000.

threads

an integer specifying the number of threads to use. The default value is 1.

alternative

a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater", or "less".

...

additional arguments that can be passed to the function.

Value

A data frame containing the results of the hypothesis test, with the following columns:

flag

a flagging indicator where 1 means statistically higher than expected and -1 means statistically lower than expected.

p-value

the p-value of the hypothesis test.

stat

the test statistic.

Std.Error

The standard error of the provider effect estimate, included only when test = "wald".

Details

By default, the function uses the "exact.poisbinom" method. The wald test is invalid for extreme providers (i.e. when provider effect goes to infinity). For the score test, consider that when the number of tested providers is large, refitting the models to get the restricted MLEs will take a long time. Therefore, we use unrestricted MLEs to replace the restricted MLEs during the testing procedure by default. However, the user can specify score_modified = FALSE to perform a standard score test.

References

Wu, W, Yang, Y, Kang, J, He, K. (2022) Improving large-scale estimation and inference for profiling health care providers. Statistics in Medicine, 41(15): 2840-2853.

Examples

data(ExampleDataBinary)
outcome = ExampleDataBinary$Y
covar = ExampleDataBinary$Z
ID = ExampleDataBinary$ID
fit_fe <- logis_fe(Y = outcome, Z = covar, ID = ID, message = FALSE)
test(fit_fe, test = "score")
#>     flag      p value        stat
#> 1      0 2.368547e-01 -1.18288504
#> 2      0 5.275655e-01  0.63172674
#> 3      0 2.407223e-01 -1.17318325
#> 4      0 4.103448e-01  0.82328704
#> 5      0 8.518166e-01 -0.18680109
#> 6      0 7.340842e-01  0.33969768
#> 7      0 8.678381e-01  0.16640513
#> 8      0 8.234055e-01 -0.22316704
#> 9     -1 1.175431e-03 -3.24477586
#> 10     1 8.789591e-03  2.62013132
#> 11     0 9.760882e-02  1.65655822
#> 12     0 8.217279e-01  0.22532311
#> 13     0 2.693589e-01  1.10454014
#> 14     0 1.351957e-01 -1.49392322
#> 15     0 8.366228e-01  0.20621522
#> 16    -1 3.074323e-02 -2.16038050
#> 17     0 6.828884e-02  1.82309611
#> 18    -1 3.505997e-02 -2.10766509
#> 19     0 5.591967e-01  0.58403508
#> 20     0 9.898703e-01  0.01269602
#> 21     0 6.651296e-02  1.83495095
#> 22     1 2.501334e-04  3.66212330
#> 23     0 7.014939e-01  0.38330466
#> 24     0 9.362285e-02 -1.67658915
#> 25     0 6.129192e-02 -1.87138369
#> 26     0 9.866786e-01 -0.01669667
#> 27     0 3.722253e-01  0.89231272
#> 28     0 4.530424e-01 -0.75035249
#> 29     0 5.111253e-01  0.65708654
#> 30     1 4.913776e-02  1.96739439
#> 31     0 6.844535e-01  0.40639348
#> 32     0 4.298946e-01 -0.78937208
#> 33     0 5.470421e-01 -0.60219837
#> 34     0 5.637157e-01 -0.57733124
#> 35     0 1.028324e-01  1.63127460
#> 36     0 7.631298e-02  1.77249162
#> 37     0 1.763862e-01 -1.35196586
#> 38     0 6.671747e-02 -1.83357257
#> 39     0 3.840996e-01 -0.87036748
#> 40    -1 3.642754e-04 -3.56469798
#> 41    -1 3.438016e-03 -2.92559189
#> 42     0 8.481478e-01  0.19148218
#> 43     0 1.386651e-01  1.48078024
#> 44     0 1.023298e-01 -1.63366235
#> 45     1 4.032529e-02  2.05040122
#> 46     0 9.503849e-02 -1.66939820
#> 47     0 3.003560e-01  1.03567016
#> 48    -1 2.077418e-02 -2.31206350
#> 49    -1 9.985059e-03 -2.57634628
#> 50    -1 2.191895e-04 -3.69580693
#> 51     0 7.801087e-02 -1.76234592
#> 52     0 2.975269e-01 -1.04175157
#> 53     0 6.478207e-01  0.45679184
#> 54     0 8.422148e-01  0.19906131
#> 55     0 1.239934e-01 -1.53822567
#> 56     0 3.609434e-01  0.91356884
#> 57     0 3.786168e-01  0.88044773
#> 58     0 2.121593e-01 -1.24764986
#> 59     0 7.466727e-01  0.32302978
#> 60     0 4.544028e-01 -0.74809513
#> 61     1 5.356552e-04  3.46226211
#> 62     0 6.611430e-02  1.83764793
#> 63     0 3.474364e-01  0.93957353
#> 64     0 3.852536e-01  0.86825714
#> 65     0 9.967024e-02 -1.64645442
#> 66     0 9.901912e-01 -0.01229387
#> 67     0 2.857652e-01 -1.06745767
#> 68     0 8.733720e-01  0.15937680
#> 69     0 3.132875e-01 -1.00834791
#> 70     0 2.083547e-01 -1.25810237
#> 71     0 5.543279e-01 -0.59128739
#> 72    -1 1.522301e-03 -3.17039826
#> 73     0 8.844973e-02 -1.70363269
#> 74    -1 1.448525e-02 -2.44499902
#> 75     1 4.095939e-03  2.87067690
#> 76     0 3.707370e-01  0.89509379
#> 77     0 9.217127e-02 -1.68405380
#> 78     0 3.036570e-01  1.02862272
#> 79     0 8.790315e-01 -0.15219705
#> 80    -1 5.365323e-03 -2.78424050
#> 81     0 1.431811e-01 -1.46404714
#> 82     0 4.143958e-01  0.81618239
#> 83     0 1.775808e-01  1.34824130
#> 84     0 1.390201e-01  1.47944991
#> 85     0 8.999547e-02  1.69542161
#> 86    -1 3.262394e-02 -2.13668106
#> 87     0 5.039333e-01 -0.66831374
#> 88     0 3.066655e-01  1.02224379
#> 89     0 2.873969e-01 -1.06384938
#> 90     0 2.530347e-01  1.14300926
#> 91     0 3.377431e-01  0.95863405
#> 92     0 9.162883e-01  0.10511032
#> 93     0 1.202064e-01 -1.55390796
#> 94    -1 6.235120e-03 -2.73515348
#> 95     0 1.165077e-01  1.56960197
#> 96     1 8.359125e-06  4.45577351
#> 97     0 2.953818e-01  1.04638830
#> 98     0 4.866249e-01 -0.69568697
#> 99     0 6.223881e-01 -0.49246856
#> 100    0 2.861710e-01  1.06655902